Sequence analysis of area of interest entries

Analyses the sequence of entries into defined AOI regions across trials. Should be used with fixation data.

Usage

AOI_seq(
  data,
  AOIs,
  AOI_names = NULL,
  sample_rate = NULL,
  long = TRUE,
  participant_ID = "participant_ID"
)

Arguments

data

A dataframe with fixation data (from fixation_dispersion). Either single or multi participant data

AOIs

A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).

AOI_names

An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.

sample_rate

Optional sample rate of the eye-tracker (Hz) for use with raw_data. If not supplied, the sample rate will be estimated from the time column and the number of samples.

long

Whether to return the AOI fixations in long or wide format. Defaults to long

participant_ID

the variable that determines the participant identifier. If no column present, assumes a single participant

Value

a long format dataframe containing the sequence of entries into AOIs on each trial

Examples

fix_d <- fixation_dispersion(example_raw_WM)
AOI_seq(fix_d, AOIs = AOIs_WM)
#>     trial AOI entry_n
#> 1       1   3       1
#> 2       1   4       2
#> 3       1   3       3
#> 4       1   2       4
#> 5       1   4       5
#> 6       1   2       6
#> 7       2   4       1
#> 8       2   3       2
#> 9       2   1       3
#> 10      2   2       4
#> 11      2   4       5
#> 12      3   2       1
#> 13      3   4       2
#> 14      3   3       3
#> 15      3   1       4
#> 16      3   3       5
#> 17      4   1       1
#> 18      4   2       2
#> 19      4   4       3
#> 20      4   3       4
#> 21      4   1       5
#> 22      4   4       6
#> 23      5   1       1
#> 24      5   2       2
#> 25      5   1       3
#> 26      5   2       4
#> 27      5   4       5
#> 28      6   2       1
#> 29      6   4       2
#> 30      6   2       3
#> 31      6   1       4
#> 32      7   2       1
#> 33      7   4       2
#> 34      7   2       3
#> 35      7   3       4
#> 36      7   4       5
#> 37      7   2       6
#> 38      7   1       7
#> 39      8   2       1
#> 40      8   4       2
#> 41      8   1       3
#> 42      9   3       1
#> 43      9   1       2
#> 44      9   3       3
#> 45      9   1       4
#> 46      9   2       5
#> 47     10   2       1
#> 48     10   1       2
#> 49     10   3       3
#> 50     11   4       1
#> 51     11   3       2
#> 52     11   1       3
#> 53     11   3       4
#> 54     11   4       5
#> 55     12   2       1
#> 56     12   4       2
#> 57     12   2       3
#> 58     12   1       4
#> 59     12   3       5
#> 60     13   3       1
#> 61     13   2       2
#> 62     13   1       3
#> 63     13   3       4
#> 64     14   1       1
#> 65     14   3       2
#> 66     14   1       3
#> 67     14   3       4
#> 68     14   1       5
#> 69     15   4       1
#> 70     15   2       2
#> 71     15   1       3
#> 72     15   3       4
#> 73     16   1       1
#> 74     16   3       2
#> 75     16   4       3
#> 76     16   2       4
#> 77     17   2       1
#> 78     17   1       2
#> 79     17   3       3
#> 80     17   4       4
#> 81     17   1       5
#> 82     17   2       6
#> 83     18   1       1
#> 84     18   3       2
#> 85     18   4       3
#> 86     19   4       1
#> 87     19   2       2
#> 88     19   1       3
#> 89     20   2       1
#> 90     20   4       2
#> 91     20   1       3
#> 92     20   3       4
#> 93     21   4       1
#> 94     21   2       2
#> 95     21   4       3
#> 96     21   3       4
#> 97     22   4       1
#> 98     22   2       2
#> 99     22   4       3
#> 100    22   2       4
#> 101    22   3       5
#> 102    23   3       1
#> 103    23   1       2
#> 104    23   3       3
#> 105    23   4       4
#> 106    23   2       5
#> 107    24   2       1
#> 108    24   4       2
#> 109    24   2       3
#> 110    24   1       4
#> 111    25   1       1
#> 112    25   1       2
#> 113    25   3       3
#> 114    25   2       4
#> 115    26   2       1
#> 116    26   4       2
#> 117    26   2       3
#> 118    27   2       1
#> 119    27   4       2
#> 120    27   2       3
#> 121    27   4       4
#> 122    27   2       5
#> 123    27   1       6
#> 124    28   2       1
#> 125    28   4       2
#> 126    29   4       1
#> 127    29   2       2
#> 128    29   4       3
#> 129    29   2       4
#> 130    30   2       1
#> 131    30   4       2
#> 132    30   2       3
#> 133    30   4       4
#> 134    30   1       5
#> 135    30   4       6
#> 136    31   4       1
#> 137    31   3       2
#> 138    31   2       3
#> 139    31   3       4
#> 140    31   1       5
#> 141    31   4       6
#> 142    32   4       1
#> 143    32   2       2
#> 144    32   4       3
#> 145    32   2       4
#> 146    33   2       1
#> 147    33   4       2
#> 148    33   1       3
#> 149    33   3       4
#> 150    34   3       1
#> 151    34   1       2
#> 152    34   2       3
#> 153    34   4       4
#> 154    35   2       1
#> 155    36   2       1
#> 156    36   4       2
#> 157    36   3       3
#> 158    37   2       1
#> 159    37   4       2
#> 160    37   1       3
#> 161    37   3       4
#> 162    37   2       5
#> 163    38   2       1
#> 164    38   4       2
#> 165    38   2       3
#> 166    38   4       4
#> 167    38   1       5
#> 168    38   3       6
#> 169    39   3       1
#> 170    39   2       2
#> 171    39   1       3
#> 172    39   2       4
#> 173    39   4       5
#> 174    39   3       6
#> 175    40   4       1
#> 176    40   2       2
#> 177    40   1       3
#> 178    40   3       4
#> 179    40   4       5
#> 180    40   2       6
#> 181    41   1       1
#> 182    41   3       2
#> 183    41   4       3
#> 184    42   4       1
#> 185    42   2       2
#> 186    42   1       3
#> 187    43   3       1
#> 188    43   2       2
#> 189    43   1       3
#> 190    43   3       4
#> 191    43   1       5
#> 192    44   2       1
#> 193    44   4       2
#> 194    44   2       3
#> 195    44   4       4
#> 196    44   2       5
#> 197    44   4       6
#> 198    45   2       1
#> 199    45   4       2
#> 200    45   2       3
#> 201    45   4       4
#> 202    45   1       5
#> 203    46   2       1
#> 204    46   4       2
#> 205    46   2       3
#> 206    46   4       4
#> 207    46   1       5
#> 208    47   2       1
#> 209    47   4       2
#> 210    47   2       3
#> 211    47   1       4
#> 212    47   2       5
#> 213    47   3       6
#> 214    48   2       1
#> 215    48   4       2
#> 216    48   2       3
#> 217    48   4       4
#> 218    49   4       1
#> 219    49   2       2
#> 220    49   4       3
#> 221    49   2       4
#> 222    49   1       5
#> 223    50   4       1
#> 224    50   3       2
#> 225    50   1       3
#> 226    50   2       4
#> 227    50   4       5
#> 228    51   1       1
#> 229    51   2       2
#> 230    51   1       3
#> 231    52   4       1
#> 232    52   2       2
#> 233    52   4       3
#> 234    52   2       4
#> 235    52   3       5
#> 236    53   1       1
#> 237    53   2       2
#> 238    53   4       3
#> 239    54   3       1
#> 240    54   1       2
#> 241    54   2       3
#> 242    54   4       4
#> 243    55   4       1
#> 244    55   2       2
#> 245    55   4       3
#> 246    55   2       4
#> 247    55   4       5
#> 248    56   2       1
#> 249    56   4       2
#> 250    56   1       3
#> 251    56   3       4
#> 252    57   1       1
#> 253    57   3       2
#> 254    57   4       3
#> 255    57   2       4
#> 256    58   4       1
#> 257    58   2       2
#> 258    58   4       3
#> 259    58   2       4
#> 260    58   1       5
#> 261    59   4       1
#> 262    59   2       2
#> 263    59   1       3
#> 264    59   3       4
#> 265    60   2       1
#> 266    60   4       2
#> 267    60   1       3
#> 268    60   3       4
#> 269    60   2       5
#> 270    61   2       1
#> 271    61   4       2
#> 272    61   2       3
#> 273    61   1       4
#> 274    61   3       5
#> 275    62   1       1
#> 276    63   4       1
#> 277    63   2       2
#> 278    63   4       3
#> 279    63   2       4
#> 280    63   1       5
#> 281    63   3       6
#> 282    64   2       1
#> 283    64   4       2
#> 284    64   2       3
#> 285    64   1       4
#> 286    64   3       5
#> 287    65   1       1
#> 288    65   3       2
#> 289    65   1       3
#> 290    65   3       4
#> 291    65   4       5
#> 292    65   2       6
#> 293    66   1       1
#> 294    66   2       2
#> 295    66   1       3
#> 296    66   3       4
#> 297    66   4       5
#> 298    66   2       6
#> 299    67   2       1
#> 300    67   4       2
#> 301    67   1       3
#> 302    68   2       1
#> 303    68   4       2
#> 304    68   2       3
#> 305    68   4       4
#> 306    68   1       5
#> 307    69   4       1
#> 308    69   3       2
#> 309    69   1       3
#> 310    69   2       4
#> 311    70   1       1
#> 312    70   2       2
#> 313    70   4       3
#> 314    70   3       4
#> 315    71   4       1
#> 316    71   2       2
#> 317    71   4       3
#> 318    71   2       4
#> 319    71   1       5
#> 320    72   4       1
#> 321    72   2       2
#> 322    72   1       3
#> 323    72   4       4
#> 324    73   1       1
#> 325    73   3       2
#> 326    73   4       3
#> 327    73   2       4
#> 328    73   4       5
#> 329    73   1       6
#> 330    74   3       1
#> 331    74   1       2
#> 332    74   2       3
#> 333    75   2       1
#> 334    75   4       2
#> 335    75   3       3
#> 336    75   2       4
#> 337    76   4       1
#> 338    76   2       2
#> 339    76   1       3
#> 340    76   3       4
#> 341    76   4       5
#> 342    77   3       1
#> 343    77   1       2
#> 344    77   2       3
#> 345    77   4       4
#> 346    78   4       1
#> 347    78   2       2
#> 348    78   1       3
#> 349    79   1       1
#> 350    79   2       2
#> 351    79   4       3
#> 352    79   3       4
#> 353    79   1       5
#> 354    80   1       1
#> 355    80   3       2
#> 356    80   4       3
#> 357    80   2       4
#> 358    81   2       1
#> 359    81   4       2
#> 360    81   2       3
#> 361    81   4       4
#> 362    81   1       5
#> 363    82   1       1
#> 364    82   2       2
#> 365    82   4       3
#> 366    83   3       1
#> 367    83   2       2
#> 368    83   4       3
#> 369    83   1       4
#> 370    84   2       1
#> 371    84   4       2
#> 372    84   1       3
#> 373    85   2       1
#> 374    85   4       2
#> 375    85   1       3
#> 376    86   3       1
#> 377    86   1       2
#> 378    87   4       1
#> 379    87   2       2
#> 380    87   4       3
#> 381    87   4       4
#> 382    87   2       5
#> 383    88   4       1
#> 384    88   2       2
#> 385    88   4       3
#> 386    88   2       4
#> 387    88   1       5
#> 388    89   3       1
#> 389    89   1       2
#> 390    89   4       3
#> 391    90   2       1
#> 392    90   4       2
#> 393    90   2       3
#> 394    90   4       4
#> 395    90   1       5
#> 396    91   3       1
#> 397    91   1       2
#> 398    92   2       1
#> 399    92   4       2
#> 400    92   2       3
#> 401    92   1       4
#> 402    92   3       5
#> 403    93   1       1
#> 404    93   4       2
#> 405    93   3       3
#> 406    93   4       4
#> 407    93   2       5
#> 408    94   2       1
#> 409    94   4       2
#> 410    94   1       3
#> 411    94   3       4
#> 412    95   2       1
#> 413    95   4       2
#> 414    95   1       3
#> 415    96   3       1
#> 416    96   1       2
#> 417    96   4       3
#> 418    97   3       1
#> 419    97   1       2
#> 420    97   3       3
#> 421    97   3       4
#> 422    98   1       1
#> 423    98   3       2
#> 424    98   4       3
#> 425    98   2       4
#> 426    99   4       1
#> 427    99   3       2
#> 428    99   2       3
#> 429    99   3       4
#> 430   100   1       1