Fixation detection using a velocity threshold identification method

Determine fixations by assessing the velocity of eye-movements, using a method that is similar to that proposed by Salvucci & Goldberg (2000). Applies the algorithm used in VTI_saccade and removes the identified saccades before assessing whether separated fixations are outside of the dispersion tolerance. If they are outside of this tolerance, the fixation is treated as a new fixation regardless of the length of saccade separating them. Compared to fixation_dispersion(), fixation_VTI() is more conservative in determining a fixation as smaller saccades are discounted and the resulting data is treated as a continued fixation (assuming it is within the pixel tolerance set by disp_tol). Returns a summary of the fixations found per trial, including start and end coordinates, timing, duration, mean velocity, and peak velocity.

Usage

fixation_VTI(
  data,
  threshold = 100,
  min_dur = 150,
  min_dur_sac = 20,
  disp_tol = 100,
  smooth = FALSE,
  progress = TRUE
)

Arguments

data

A dataframe with raw data (pID, time, x, y, trial), the standardised raw data form for eyetools

threshold

velocity threshold (degrees of VA / sec) to be used for identifying saccades.

min_dur

Minimum duration (in milliseconds) of period over which fixations are assessed

min_dur_sac

Minimum duration (in milliseconds) for saccades to be determined

disp_tol

Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period

smooth

include a call to eyetools::smoother on each trial

progress

Display a progress bar

Value

a dataframe containing each detected fixation by trial, with mean x/y position in pixel, start and end times, and duration.

Details

Analyses data separately for each unique combination of values in pID and trial.

References

Salvucci, D. D., & Goldberg, J. H. (2000). Identifying fixations and saccades in eye-tracking protocols. Proceedings of the Symposium on Eye Tracking Research & Applications - ETRA '00, 71–78.

Examples

# \donttest{
data <- combine_eyes(HCL)
data <- interpolate(data)
fixation_VTI(data)
#>     pID trialNumber fix_n start   end duration         x        y min_dur
#> 1   118           1     1     0   177      177  960.6437 812.8221     150
#> 2   118           1     2   203   430      227  958.2697 583.3656     150
#> 3   118           1     3   447   747      300  960.3778 475.9596     150
#> 4   118           1     4   783  1190      407  543.1279 834.5434     150
#> 5   118           1     5  1220  1730      510  939.8145 500.3046     150
#> 6   118           1     6  1753  2190      437  964.9265 168.6218     150
#> 7   118           1     7  2246  2686      440 1538.1665 827.9873     150
#> 8   118           1     8  2753  2916      163  966.5097 351.1727     150
#> 9   118           1     9  2933  3133      200  949.4499 449.1512     150
#> 10  118           1    10  3126  3676      550  945.8062 118.8186     150
#> 11  118           1    11  3736  4219      483 1517.4546 807.8438     150
#> 12  118           1    12  4283  4473      190  973.7590 337.2340     150
#> 13  118           1    13  4496  5373      877  962.4422 485.2484     150
#> 14  118           1    14  5343  6822     1479  512.2158 853.7036     150
#> 15  118           1    15  6812  6999      187  919.8259 641.8424     150
#> 16  118           1    16  7136  7596      460  966.9237 621.3341     150
#> 17  118           1    17  7646  8222      576  510.5845 825.7554     150
#> 18  118           1    18  8229  9289     1060  508.0056 858.4632     150
#> 19  118           1    19  9342 10059      717  939.2560 463.0083     150
#> 20  118           1    20 10095 10465      370  945.8363 131.5389     150
#> 21  118           1    21 10499 10699      200  552.2622 804.5069     150
#> 22  118           1    22 10749 11115      366  940.8964 462.0165     150
#> 23  118           1    23 11142 11902      760  951.3329 280.4016     150
#> 24  118           1    24 11928 12135      207  958.5555 430.8222     150
#> 25  118           1    25 12198 12968      770 1521.8775 829.2840     150
#> 26  118           1    26 13032 13431      399  951.5551 442.6154     150
#> 27  118           2     1     0   173      173  958.3036 777.7691     150
#> 28  118           2     2   193   380      187  974.5044 631.8866     150
#> 29  118           2     3   376  1196      820 1506.1300 812.5378     150
#> 30  118           2     4  1189  1426      237  950.4405 395.8372     150
#> 31  118           2     5  1449  1796      347  943.4877 146.6723     150
#> 32  118           2     6  1819  2009      190  935.8919 483.3098     150
#> 33  118           2     7  2029  2619      590  493.0987 820.6721     150
#> 34  118           2     8  2689  3102      413  941.2462 172.4029     150
#> 35  118           2     9  3132  3539      407  954.6329 461.6827     150
#> 36  118           2    10  3566  3919      353  946.3553 180.3700     150
#> 37  118           2    11  3939  4266      327  937.1356 289.0297     150
#> 38  118           2    12  4286  5136      850  937.7621 207.9633     150
#> 39  118           2    13  5399  5759      360  503.2766 825.4297     150
#> 40  118           2    14  5895  6229      334  948.9428 156.7683     150
#> 41  118           2    15  6289  6745      456  477.9152 805.5287     150
#> 42  118           2    16  6799  7202      403  953.1376 473.4159     150
#> 43  118           3     1     0   220      220        NA       NA     150
#> 44  118           3     2   403   683      280  962.6799 456.5565     150
#> 45  118           3     3   706   933      227  973.7131 177.9272     150
#> 46  118           3     4   993  1200      207 1501.6394 843.7145     150
#> 47  118           3     5  1260  1826      566  496.7830 834.2048     150
#> 48  118           3     6  1876  2270      394  949.6194 464.7100     150
#> 49  118           3     7  2290  2619      329  970.7820 184.9187     150
#> 50  118           3     8  2656  2989      333  951.6396 477.7824     150
#> 51  118           3     9  3056  3553      497 1518.9085 851.4880     150
#> 52  118           3    10  3586  3743      157 1005.1878 577.4765     150
#> 53  118           3    11  3873  4113      240  962.3067 471.0581     150
#> 54  118           3    12  4133  4543      410  947.0560 301.5257     150
#> 55  118           3    13  4566  4759      193  942.4467 443.0341     150
#> 56  118           3    14  4779  5263      484  951.4985 203.0580     150
#> 57  118           4     1     0   160      160  956.9020 813.5203     150
#> 58  118           4     2   193   343      150 1406.2372 800.3531     150
#> 59  118           4     3   360   540      180 1507.5290 801.4327     150
#> 60  118           4     4   560   913      353  994.3665 428.3371     150
#> 61  118           4     5  1103  1536      433  506.2403 840.4780     150
#> 62  118           4     6  1650  2383      733 1514.6499 847.2079     150
#> 63  118           4     7  2450  3346      896  526.0427 826.3692     150
#> 64  118           4     8  3406  3689      283  981.7425 439.5356     150
#> 65  118           4     9  3716  4153      437  958.8945 153.7166     150
#> 66  118           4    10  4219  4709      490 1478.6775 814.2137     150
#> 67  118           4    11  4819  5452      633  513.9950 844.8581     150
#> 68  118           4    12  5516  5839      323  925.1128 179.0145     150
#> 69  118           4    13  5856  6166      310  955.6167 474.1818     150
#> 70  118           4    14  6216  8059     1843 1527.7158 852.0435     150
#> 71  118           4    15  8022  8609      587  539.4569 898.9155     150
#> 72  118           4    16  8642  8912      270 1502.5136 866.6847     150
#> 73  118           4    17  8975  9182      207  942.5158 390.7862     150
#> 74  118           4    18  9199  9712      513  949.4149 261.9195     150
#> 75  118           5     1   196   433      237  640.0432 805.7758     150
#> 76  118           5     2   476   783      307  918.5103 452.4816     150
#> 77  118           5     3   763   973      210  676.9740 803.1841     150
#> 78  118           5     4  1099  1659      560 1510.1178 843.3821     150
#> 79  118           5     5  1719  1959      240  981.3653 431.4367     150
#> 80  118           5     6  1986  2606      620  954.0925 247.7249     150
#> 81  118           5     7  2659  2992      333 1528.8640 795.0453     150
#> 82  118           5     8  3049  3679      630  954.3285 299.2864     150
#> 83  118           5     9  3742  3932      190 1513.6826 806.9528     150
#> 84  118           5    10  3972  4266      294  956.1005 161.4779     150
#> 85  118           6     1     0   173      173  967.6325 804.3820     150
#> 86  118           6     2   357   577      220  945.8363 433.1657     150
#> 87  118           6     3   633  1210      577 1502.6897 820.9648     150
#> 88  118           6     4  1167  1523      356  541.0264 812.6489     150
#> 89  118           6     5  1720  1970      250  942.4061 155.3532     150
#> 90  118           6     6  1986  2373      387  946.8782 467.5378     150
#> 91  118           6     7  2396  2823      427  947.6286 166.4201     150
#> 92  118           6     8  2836  3326      490  936.0810 443.3592     150
#> 93  118           6     9  3426  3756      330 1529.1779 807.2287     150
#> 94  118           6    10  3826  4283      457  493.4553 816.6774     150
#> 95  118           6    11  4333  4663      330  959.4518 504.3423     150
#> 96  118           6    12  4679  5159      480  953.6388 260.8693     150
#> 97  119           1     1     0   223      223  972.1103 765.7244     150
#> 98  119           1     2   240   419      179  936.9682 643.4583     150
#> 99  119           1     3   439  1006      567  971.1363 421.9248     150
#> 100 119           1     4   989  1406      417  412.8899 751.2055     150
#> 101 119           1     5  1399  1573      174  804.7938 509.7751     150
#> 102 119           1     6  1589  1856      267  971.0424 304.7891     150
#> 103 119           1     7  1889  2076      187  981.7317 582.0018     150
#> 104 119           1     8  2102  3026      924 1500.1104 765.8978     150
#> 105 119           1     9  3089  3296      207  445.5798 735.4684     150
#> 106 119           1    10  3316  4162      846  963.7025 480.1758     150
#> 107 119           1    11  4185  4659      474  462.4674 760.9068     150
#> 108 119           1    12  4705  5059      354 1466.6914 752.2349     150
#> 109 119           1    13  5198  7351     2153  438.1199 764.9565     150
#> 110 119           1    14  7371  7598      227 1480.2390 767.7494     150
#> 111 119           1    15  7618  7895      277 1585.6331 742.2156     150
#> 112 119           1    16  7941  8444      503  996.0639 325.5463     150
#> 113 119           1    17  8501  8851      350  343.2499 673.4041     150
#> 114 119           1    18  8888  9108      220  968.4899 314.5223     150
#> 115 119           1    19  9154  9601      447  451.1232 775.7702     150
#> 116 119           1    20  9641  9981      340  914.0899 497.4818     150
#> 117 119           2     1     0   160      160  964.4077 800.0306     150
#> 118 119           2     2   200   563      363  489.4246 748.2445     150
#> 119 119           2     3   626   783      157 1432.7215 766.1448     150
#> 120 119           2     4   800  1086      286 1546.5018 746.7085     150
#> 121 119           2     5  1273  1493      220  997.7612 399.4104     150
#> 122 119           2     6  1513  2086      573  950.6969 208.1089     150
#> 123 119           2     7  2106  2963      857  979.5058 415.4418     150
#> 124 119           2     8  3006  3216      210  934.7593 642.2169     150
#> 125 119           2     9  3249  3563      314  512.3586 766.2387     150
#> 126 119           2    10  3606  3996      390  959.0141 278.7691     150
#> 127 119           2    11  4026  4212      186  792.7098 450.0493     150
#> 128 119           2    12  4256  4462      206  518.6455 786.5630     150
#> 129 119           2    13  4526  5066      540 1472.5163 764.4277     150
#> 130 119           3     1     0   200      200  981.6611 730.4572     150
#> 131 119           3     2   194   717      523  481.2994 752.0970     150
#> 132 119           3     3   727  1167      440 1463.6185 711.6945     150
#> 133 119           3     4  1190  1650      460  968.9511 420.4636     150
#> 134 119           3     5  1670  2033      363  988.2300 123.7073     150
#> 135 119           3     6  2413  3106      693 1516.4708 753.9228     150
#> 136 119           3     7  3183  3923      740  446.6575 755.3015     150
#> 137 119           3     8  3996  4753      757 1499.3788 763.7812     150
#> 138 119           3     9  4763  5013      250 1386.0553 699.7711     150
#> 139 119           3    10  5086  5686      600  400.0376 714.8983     150
#> 140 119           3    11  5749  6782     1033  980.9620 436.1020     150
#> 141 119           3    12  6802  7132      330  966.1208 300.9559     150
#> 142 119           4     1     0   253      253  971.3683 746.0566     150
#> 143 119           4     2   270  1217      947  427.5659 747.1888     150
#> 144 119           4     3  1446  1873      427 1441.6189 745.8055     150
#> 145 119           4     4  1960  2523      563  537.3871 766.3947     150
#> 146 119           4     5  2546  3000      454  395.6776 748.1303     150
#> 147 119           4     6  3036  3253      217  908.1463 479.9049     150
#> 148 119           4     7  3299  4139      840  488.8203 746.7495     150
#> 149 119           4     8  4209  5412     1203 1455.0616 710.4195     150
#> 150 119           4     9  5482  6059      577  482.4678 764.8231     150
#> 151 119           4    10  6232  6619      387 1528.0690 738.6647     150
#> 152 119           4    11  6692  7352      660  450.2031 774.6959     150
#> 153 119           4    12  7559  7742      183  965.2065 383.8848     150
#> 154 119           4    13  7772  8178      406  969.9090 144.8427     150
#> 155 119           4    14  8232  8598      366  950.2556 536.2692     150
#> 156 119           5     1     0   210      210  976.5170 750.1115     150
#> 157 119           5     2   244   477      233  492.8740 735.6804     150
#> 158 119           5     3   497   917      420  371.6881 770.6937     150
#> 159 119           5     4   937  1103      166  508.2954 782.8994     150
#> 160 119           5     5  1100  2033      933 1486.8775 762.0913     150
#> 161 119           5     6  2143  2726      583 1607.9763 776.6748     150
#> 162 119           5     7  2886  3386      500  457.2755 767.4715     150
#> 163 119           5     8  3393  3989      596 1509.5685 747.9117     150
#> 164 119           5     9  4009  4496      487 1431.5261 767.8143     150
#> 165 119           5    10  4579  4826      247  493.5553 762.4451     150
#> 166 119           5    11  4846  5289      443  349.1937 779.2816     150
#> 167 119           5    12  5286  5513      227  521.5009 751.7369     150
#> 168 119           5    13  5599  6596      997 1502.7039 779.1490     150
#> 169 119           5    14  6689  7402      713  524.1221 763.3493     150
#> 170 119           5    15  7439  8209      770  966.2373 418.0393     150
#> 171 119           5    16  8242  9005      763  976.8946 150.0364     150
#> 172 119           5    17  9089  9562      473  471.4173 707.9541     150
#> 173 119           5    18  9598 10148      550  958.5243 327.3672     150
#> 174 119           5    19 10168 10425      257  980.3276 178.7893     150
#> 175 119           5    20 10595 11755     1160  392.8766 759.6779     150
#> 176 119           5    21 11761 12374      613  927.5780 321.4340     150
#> 177 119           6     1     0   236      236  963.4609 755.3811     150
#> 178 119           6     2   256   436      180  962.9117 617.2494     150
#> 179 119           6     3   626  2266     1640  413.3380 790.3672     150
#> 180 119           6     4  2276  2433      157  591.3804 756.9897     150
#> 181 119           6     5  2453  3049      596  468.5629 739.4379     150
#> 182 119           6     6  3179  4086      907 1534.3101 784.0435     150
#> 183 119           6     7  4269  4605      336  991.7024 481.1540     150
#> 184 119           6     8  4656  5245      589  950.8301 201.0391     150
#> 185 119           6     9  5269  6775     1506  980.3141 430.5043     150
#> 186 119           6    10  6885  7555      670 1586.3520 735.5883     150
#> 187 119           6    11  7608  7828      220 1043.4398 360.0337     150
#> 188 119           6    12  7848  9161     1313  952.2215 181.7543     150
#> 189 119           6    13  9181  9358      177  948.7052 266.9317     150
#> 190 119           6    14  9428 10088      660 1536.3246 741.4289     150
#> 191 119           6    15 10098 10278      180 1418.9720 760.8927     150
#> 192 119           6    16 10351 11014      663  464.2243 786.3815     150
#> 193 119           6    17 11088 11278      190 1333.9099 779.3765     150
#> 194 119           6    18 11297 11577      280 1524.7641 751.3519     150
#> 195 119           6    19 11647 11994      347  439.4383 728.0833     150
#> 196 119           6    20 12071 12534      463  988.9682 311.6586     150
#>     disp_tol
#> 1        100
#> 2        100
#> 3        100
#> 4        100
#> 5        100
#> 6        100
#> 7        100
#> 8        100
#> 9        100
#> 10       100
#> 11       100
#> 12       100
#> 13       100
#> 14       100
#> 15       100
#> 16       100
#> 17       100
#> 18       100
#> 19       100
#> 20       100
#> 21       100
#> 22       100
#> 23       100
#> 24       100
#> 25       100
#> 26       100
#> 27       100
#> 28       100
#> 29       100
#> 30       100
#> 31       100
#> 32       100
#> 33       100
#> 34       100
#> 35       100
#> 36       100
#> 37       100
#> 38       100
#> 39       100
#> 40       100
#> 41       100
#> 42       100
#> 43       100
#> 44       100
#> 45       100
#> 46       100
#> 47       100
#> 48       100
#> 49       100
#> 50       100
#> 51       100
#> 52       100
#> 53       100
#> 54       100
#> 55       100
#> 56       100
#> 57       100
#> 58       100
#> 59       100
#> 60       100
#> 61       100
#> 62       100
#> 63       100
#> 64       100
#> 65       100
#> 66       100
#> 67       100
#> 68       100
#> 69       100
#> 70       100
#> 71       100
#> 72       100
#> 73       100
#> 74       100
#> 75       100
#> 76       100
#> 77       100
#> 78       100
#> 79       100
#> 80       100
#> 81       100
#> 82       100
#> 83       100
#> 84       100
#> 85       100
#> 86       100
#> 87       100
#> 88       100
#> 89       100
#> 90       100
#> 91       100
#> 92       100
#> 93       100
#> 94       100
#> 95       100
#> 96       100
#> 97       100
#> 98       100
#> 99       100
#> 100      100
#> 101      100
#> 102      100
#> 103      100
#> 104      100
#> 105      100
#> 106      100
#> 107      100
#> 108      100
#> 109      100
#> 110      100
#> 111      100
#> 112      100
#> 113      100
#> 114      100
#> 115      100
#> 116      100
#> 117      100
#> 118      100
#> 119      100
#> 120      100
#> 121      100
#> 122      100
#> 123      100
#> 124      100
#> 125      100
#> 126      100
#> 127      100
#> 128      100
#> 129      100
#> 130      100
#> 131      100
#> 132      100
#> 133      100
#> 134      100
#> 135      100
#> 136      100
#> 137      100
#> 138      100
#> 139      100
#> 140      100
#> 141      100
#> 142      100
#> 143      100
#> 144      100
#> 145      100
#> 146      100
#> 147      100
#> 148      100
#> 149      100
#> 150      100
#> 151      100
#> 152      100
#> 153      100
#> 154      100
#> 155      100
#> 156      100
#> 157      100
#> 158      100
#> 159      100
#> 160      100
#> 161      100
#> 162      100
#> 163      100
#> 164      100
#> 165      100
#> 166      100
#> 167      100
#> 168      100
#> 169      100
#> 170      100
#> 171      100
#> 172      100
#> 173      100
#> 174      100
#> 175      100
#> 176      100
#> 177      100
#> 178      100
#> 179      100
#> 180      100
#> 181      100
#> 182      100
#> 183      100
#> 184      100
#> 185      100
#> 186      100
#> 187      100
#> 188      100
#> 189      100
#> 190      100
#> 191      100
#> 192      100
#> 193      100
#> 194      100
#> 195      100
#> 196      100
# }