Fixation detection using a dispersion method

Detects fixations by assessing dispersion of the eye position, using a method that is similar to that proposed by Salvucci & Goldberg (1996). Evaluates the maximum dispersion (distance) between x/y coordinates across a window of data. Looks for sufficient periods in which this maximum dispersion is below the specified dispersion tolerance. NAs are considered breaks in the data and are not permitted within a valid fixation period. Runs the interpolation algorithm by default to fix small breaks in the data.

Usage

fixation_dispersion(
  data,
  min_dur = 150,
  disp_tol = 100,
  run_interp = TRUE,
  NA_tol = 0.25,
  progress = TRUE
)

Arguments

data

A dataframe with raw data (time, x, y, trial) for one participant (the standardised raw data form for eyetools)

min_dur

Minimum duration (in milliseconds) of period over which fixations are assessed

disp_tol

Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period

run_interp

include a call to eyetools::interpolate on each trial

NA_tol

the proportion of NAs tolerated within any window of samples that is evaluated as a fixation

progress

Display a progress bar

Value

a dataframe containing each detected fixation by trial, with mean x/y position in pixel, start and end times, and duration.

References

Salvucci, D. D., & Goldberg, J. H. (2000). Identifying fixations and saccades in eye-tracking protocols. Proceedings of the Symposium on Eye Tracking Research & Applications - ETRA ’00, 71–78.

Examples

fixation_dispersion(example_raw_fix, disp_tol = 150)
#>     trial fix_n start   end duration    x    y prop_NA min_dur disp_tol
#> 1       1     1    83   233      150 1210  438   0.000     150      150
#> 2       1     2   250   510      260  768  575   0.000     150      150
#> 3       1     3   667   903      236  960  916   0.000     150      150
#> 4       1     4   923  1110      187  977  434   0.000     150      150
#> 5       1     5  1113  1357      244 1182  566   0.000     150      150
#> 6       1     6  1360  1577      217 1212  606   0.000     150      150
#> 7       1     7  1593  1766      173  718  631   0.000     150      150
#> 8       1     8  1773  2026      253  934  906   0.000     150      150
#> 9       1     9  2030  2770      740  940  936   0.000     150      150
#> 10      1    10  2783  2946      163 1208  688   0.000     150      150
#> 11      1    11  2950  3330      380 1194  644   0.000     150      150
#> 12      1    12  3366  3516      150  293  651   0.000     150      150
#> 13      1    13  3520  3683      163  234  661   0.000     150      150
#> 14      1    14  3700  4023      323  737  595   0.000     150      150
#> 15      1    15  4203  4520      317 1689  654   0.000     150      150
#> 16      1    16  4536  4930      394 1227  628   0.000     150      150
#> 17      1    17  4946  5313      367 1695  659   0.000     150      150
#> 18      1    18  5326  5483      157 1226  633   0.000     150      150
#> 19      1    19  5526  5679      153  273  689   0.000     150      150
#> 20      1    20  5683  7026     1343  229  656   0.000     150      150
#> 21      1    21  7063  7329      266 1042  873   0.000     150      150
#> 22      1    22  7363  7519      156  256  691   0.000     150      150
#> 23      1    23  7523  7859      336  215  663   0.000     150      150
#> 24      2     1   320   747      427  969  960   0.000     150      150
#> 25      2     2   763  1170      407 1198  619   0.000     150      150
#> 26      2     3  1187  1417      230  750  620   0.000     150      150
#> 27      2     4  1443  1697      254  231  674   0.000     150      150
#> 28      2     5  1736  2053      317 1234  648   0.000     150      150
#> 29      2     6  2070  2246      176 1512  638   0.000     150      150
#> 30      2     7  2253  2600      347 1689  635   0.000     150      150
#> 31      2     8  2640  2993      353  706  605   0.000     150      150
#> 32      2     9  3016  3263      247  221  660   0.000     150      150
#> 33      2    10  3266  3710      444  214  651   0.000     150      150
#> 34      2    11  3750  3916      166 1115  679   0.000     150      150
#> 35      2    12  3920  5003     1083 1224  650   0.000     150      150
#> 36      2    13  5023  5579      556 1695  651   0.000     150      150
#> 37      2    14  5789  6086      297  244  664   0.000     150      150
#> 38      2    15  6103  6683      580  732  612   0.000     150      150
#> 39      2    16  6703  6899      196  968  943   0.000     150      150
#> 40      2    17  6916  7466      550  715  631   0.000     150      150
#> 41      2    18  7506  7706      200  984  949   0.000     150      150
#> 42      2    19  7722  7883      161  712  646   0.000     150      150
#> 43      2    20  7909  9522     1613  210  648   0.000     150      150
#> 44      2    21  9542 10106      564  729  638   0.000     150      150
#> 45      2    22 10149 10322      173 1618  664   0.000     150      150
#> 46      2    23 10326 10915      589 1697  648   0.000     150      150
#> 47      2    24 10955 11112      157  776  608   0.000     150      150
#> 48      2    25 11115 11679      564  716  612   0.000     150      150
#> 49      3     1     3   180      177 1756  335   0.000     150      150
#> 50      3     2   207   373      166 1008  578   0.000     150      150
#> 51      3     3   387   550      163  733  617   0.000     150      150
#> 52      3     4   580   773      193  186  643   0.000     150      150
#> 53      3     5   807   963      156 1078  601   0.000     150      150
#> 54      3     6   967  1213      246  975  933   0.000     150      150
#> 55      3     7  1217  1656      439  946  941   0.000     150      150
#> 56      3     8  1676  1866      190 1221  644   0.000     150      150
#> 57      3     9  1870  2666      796 1200  607   0.000     150      150
#> 58      3    10  2710  3240      530  217  667   0.000     150      150
#> 59      3    11  3283  3453      170 1128  609   0.000     150      150
#> 60      3    12  3456  3663      207 1235  645   0.000     150      150
#> 61      3    13  3683  4300      617 1680  642   0.000     150      150
#> 62      3    14  4473  4693      220  721  613   0.000     150      150
#> 63      3    15  4696  5033      337  696  648   0.000     150      150
#> 64      3    16  5053  6089     1036 1216  620   0.000     150      150
#> 65      3    17  6129  6296      167  328  645   0.000     150      150
#> 66      3    18  6299  7109      810  228  660   0.000     150      150
#> 67      3    19  7126  7532      406  707  616   0.000     150      150
#> 68      3    20  7549  7939      390  306  662   0.000     150      150
#> 69      3    21  7962  8129      167  798  754   0.000     150      150
#> 70      3    22  8132  8989      857  703  613   0.000     150      150
#> 71      3    23  8992  9559      567  765  488   0.000     150      150
#> 72      3    24  9562  9739      177 1014  574   0.000     150      150
#> 73      3    25  9742 10172      430 1196  615   0.000     150      150
#> 74      3    26 10212 10525      313  231  676   0.000     150      150
#> 75      3    27 10529 11659     1130  212  649   0.000     150      150
#> 76      3    28 11702 12022      320 1194  646   0.000     150      150
#> 77      3    29 12049 12429      380 1672  668   0.000     150      150
#> 78      3    30 12442 13515     1073 1228  635   0.000     150      150
#> 79      3    31 13552 13998      446  950  935   0.000     150      150
#> 80      3    32 14022 14852      830 1222  627   0.000     150      150
#> 81      3    33 14872 16018     1146 1687  667   0.000     150      150
#> 82      4     1     0   217      217 1047  701   0.000     150      150
#> 83      4     2   220   380      160 1192  663   0.000     150      150
#> 84      4     3   390   557      167  998  888   0.000     150      150
#> 85      4     4   560   993      433  978  942   0.000     150      150
#> 86      4     5  1173  1487      314  981  920   0.000     150      150
#> 87      4     6  1503  2113      610 1221  635   0.000     150      150
#> 88      4     7  2133  2390      257  724  598   0.000     150      150
#> 89      4     8  2423  2583      160 1474  652   0.000     150      150
#> 90      4     9  2593  3173      580 1697  652   0.000     150      150
#> 91      4    10  3190  3776      586 1200  611   0.000     150      150
#> 92      4    11  3966  5499     1533  204  660   0.000     150      150
#> 93      4    12  5519  5726      207  622  693   0.000     150      150
#> 94      4    13  5743  6679      936  221  673   0.000     150      150
#> 95      4    14  6703  7096      393  723  626   0.000     150      150
#> 96      4    15  7116  7486      370  229  674   0.000     150      150
#> 97      4    16  7519  7906      387  993  909   0.000     150      150
#> 98      4    17  7919  8182      263  692  632   0.000     150      150
#> 99      4    18  8186  8899      713  712  604   0.000     150      150
#> 100     4    19  8919  9426      507 1198  617   0.000     150      150
#> 101     4    20  9429  9669      240 1237  656   0.000     150      150
#> 102     4    21  9686  9936      250 1665  669   0.000     150      150
#> 103     4    22  9949 10532      583 1215  628   0.000     150      150
#> 104     4    23 10726 11362      636  238  661   0.000     150      150
#> 105     4    24 11365 11659      294  226  643   0.000     150      150
#> 106     4    25 11845 12272      427 1239  653   0.000     150      150
#> 107     4    26 12292 12682      390 1681  674   0.000     150      150
#> 108     4    27 12709 12905      196 1028  576   0.000     150      150
#> 109     4    28 12919 13735      816  723  613   0.000     150      150
#> 110     4    29 13752 14089      337 1216  651   0.000     150      150
#> 111     4    30 14112 14355      243 1577  662   0.000     150      150
#> 112     4    31 14358 14875      517 1679  663   0.000     150      150
#> 113     4    32 14892 15622      730 1233  634   0.000     150      150
#> 114     5     1   177   350      173  959  572   0.000     150      150
#> 115     5     2   363  1067      704  952  948   0.000     150      150
#> 116     5     3  1080  1660      580 1205  652   0.000     150      150
#> 117     5     4  1686  3140     1454 1686  678   0.000     150      150
#> 118     6     1     0   194      194 1033  683   0.000     150      150
#> 119     6     2   197   703      506  934  927   0.000     150      150
#> 120     6     3   720   997      277 1203  649   0.000     150      150
#> 121     6     4  1087  1267      180  705  608   0.000     150      150
#> 122     6     5  1290  1857      567  220  657   0.000     150      150
#> 123     6     6  1877  2127      250  720  640   0.000     150      150
#> 124     6     7  2147  2690      543  192  662   0.000     150      150
#> 125     7     1     0   240      240  964  601   0.000     150      150
#> 126     7     2   290   447      157  952  928   0.000     150      150
#> 127     7     3   450   640      190  947  910   0.000     150      150
#> 128     7     4   707  1177      470 1228  660   0.000     150      150
#> 129     7     5  1200  1433      233 1638  676   0.000     150      150
#> 130     7     6  1437  1927      490 1686  670   0.000     150      150
#> 131     7     7  1943  2333      390 1214  664   0.000     150      150
#> 132     7     8  2357  2597      240  732  613   0.000     150      150
#> 133     7     9  2600  2783      183  674  632   0.000     150      150
#> 134     7    10  2800  3136      336  256  663   0.000     150      150
#> 135     7    11  3150  3793      643  703  617   0.000     150      150
#> 136     8     1     0   173      173 1231  634   0.000     150      150
#> 137     8     2   360   600      240  987  902   0.000     150      150
#> 138     8     3   720  1323      603 1210  636   0.000     150      150
#> 139     8     4  1343  1556      213  708  606   0.000     150      150
#> 140     8     5  1583  1800      217  258  670   0.000     150      150
#> 141     8     6  1816  2293      477  732  614   0.000     150      150
#> 142     8     7  2316  2583      267 1200  634   0.000     150      150
#> 143     8     8  2613  2960      347 1706  671   0.000     150      150
#> 144     8     9  2980  3356      376 1205  638   0.000     150      150
#> 145     9     1     0   170      170 1072  578   0.000     150      150
#> 146     9     2   183   630      447  950  917   0.000     150      150
#> 147     9     3   773   930      157  616  647   0.000     150      150
#> 148     9     4   950  2133     1183  234  678   0.000     150      150
#> 149    10     1     0   244      244 1032  619   0.000     150      150
#> 150    10     2   254   673      419  936  944   0.000     150      150
#> 151    10     3   847  1163      316 1580  675   0.000     150      150
#> 152    10     4  1167  1747      580 1679  664   0.000     150      150
#> 153    11     1     0   313      313 1013  590   0.000     150      150
#> 154    11     2   406   596      190  943  940   0.000     150      150
#> 155    11     3   610   886      276  769  604   0.000     150      150
#> 156    11     4   899  1499      600 1208  647   0.000     150      150
#> 157    11     5  1503  2223      720 1235  660   0.000     150      150
#> 158    11     6  2243  2513      270 1679  682   0.000     150      150
#> 159    11     7  2553  3456      903  714  613   0.000     150      150
#> 160    11     8  3483  4002      519  231  680   0.000     150      150
#> 161    11     9  4022  4286      264  746  599   0.000     150      150
#> 162    11    10  4289  4706      417  779  405   0.000     150      150
#> 163    11    11  4732  6166     1434  238  685   0.000     150      150
#> 164    11    12  6209  6925      716 1221  654   0.000     150      150
#> 165    12     1     0   180      180 1038  229   0.000     150      150
#> 166    12     2   197   410      213  827  561   0.000     150      150
#> 167    12     3   434  1400      966  952  941   0.000     150      150
#> 168    12     4  1440  1610      170 1654  675   0.000     150      150
#> 169    12     5  1614  2277      663 1682  651   0.000     150      150
#> 170    12     6  2293  2447      154 1255  611   0.000     150      150
#> 171    12     7  2450  4097     1647 1203  636   0.000     150      150
#> 172    12     8  4123  4463      340 1693  634   0.000     150      150
#> 173    12     9  4513  4673      160  569  634   0.000     150      150
#> 174    12    10  4826  5253      427 1190  666   0.000     150      150
#> 175    12    11  5270  5846      576  733  619   0.000     150      150
#> 176    12    12  5850  6376      526  704  633   0.000     150      150
#> 177    13     1     0   173      173 1100  680   0.000     150      150
#> 178    13     2   183   406      223  729  636   0.000     150      150
#> 179    13     3   423   770      347  984  946   0.000     150      150
#> 180    13     4   773  1126      353  943  965   0.000     150      150
#> 181    13     5  1143  1350      207 1225  650   0.000     150      150
#> 182    13     6  1373  2200      827 1671  664   0.000     150      150
#> 183    13     7  2256  2430      174  274  649   0.000     150      150
#> 184    13     8  2433  2866      433  241  689   0.000     150      150
#> 185    13     9  2893  3113      220  930  959   0.000     150      150
#> 186    13    10  3116  3426      310  957  987   0.000     150      150
#> 187    13    11  3453  3606      153 1625  691   0.000     150      150
#> 188    13    12  3609  4106      497 1686  678   0.000     150      150
#> 189    14     1    17   194      177 1004  433   0.000     150      150
#> 190    14     2   207   400      193  758  626   0.000     150      150
#> 191    14     3   414   787      373  987  923   0.000     150      150
#> 192    14     4   897  1190      293 1216  620   0.000     150      150
#> 193    14     5  1193  1670      477 1217  617   0.000     150      150
#> 194    14     6  1687  1840      153  756  610   0.000     150      150
#> 195    14     7  1857  2057      200 1228  634   0.000     150      150
#> 196    14     8  2093  2260      167  305  650   0.000     150      150
#> 197    14     9  2263  3383     1120  224  655   0.000     150      150
#> 198    14    10  3410  3767      357  767  467   0.000     150      150
#> 199    14    11  3790  4453      663 1223  637   0.000     150      150
#> 200    15     1    70   220      150 1033  552   0.000     150      150
#> 201    15     2   223   376      153 1210  631   0.000     150      150
#> 202    15     3   386   543      157  990  898   0.000     150      150
#> 203    15     4   546  1016      470  952  941   0.000     150      150
#> 204    15     5  1076  1353      277  728  598   0.000     150      150
#> 205    15     6  1370  1526      156 1024  841   0.000     150      150
#> 206    15     7  1530  1716      186  938  933   0.000     150      150
#> 207    15     8  1746  1946      200  351  672   0.000     150      150
#> 208    15     9  1956  2499      543  722  625   0.000     150      150
#> 209    15    10  2526  2983      457  246  648   0.000     150      150
#> 210    15    11  2999  3159      160  744  600   0.000     150      150
#> 211    15    12  3206  3696      490 1632  622   0.000     150      150
#> 212    15    13  3699  4279      580 1700  629   0.000     150      150
#> 213    15    14  4296  4556      260 1243  659   0.000     150      150
#> 214    15    15  4559  4966      407 1213  652   0.000     150      150
#> 215    15    16  4986  6082     1096  708  609   0.000     150      150
#> 216    16     1   154   340      186 1078  595   0.000     150      150
#> 217    16     2   354   630      276  951  938   0.000     150      150
#> 218    16     3   633  1107      474  928  934   0.000     150      150
#> 219    16     4  1110  1340      230  925  941   0.000     150      150
#> 220    16     5  1377  1660      283  294  633   0.000     150      150
#> 221    16     6  1663  2057      394  221  671   0.000     150      150
#> 222    17     1     0   220      220  933  618   0.000     150      150
#> 223    17     2   230   393      163  981  948   0.000     150      150
#> 224    17     3   397   730      333  957  948   0.000     150      150
#> 225    17     4   760  1133      373 1631  662   0.000     150      150
#> 226    17     5  1137  1447      310 1697  663   0.000     150      150
#> 227    17     6  1610  1760      150 1297  645   0.239     150      150
#> 228    17     7  1870  2043      173  263  673   0.000     150      150
#> 229    17     8  2210  2370      160  847  570   0.000     150      150
#> 230    17     9  2407  2630      223 1664  626   0.000     150      150
#> 231    17    10  2633  2996      363 1698  647   0.000     150      150
#> 232    17    11  3013  3506      493 1237  628   0.000     150      150
#> 233    17    12  3523  3853      330 1659  622   0.000     150      150
#> 234    17    13  4000  4176      176 1173  609   0.000     150      150
#> 235    17    14  4196  4480      284  732  607   0.000     150      150
#> 236    17    15  4496  5393      897 1227  634   0.000     150      150
#> 237    17    16  5413  5956      543  700  629   0.000     150      150
#> 238    18     1   190   363      173 1176  593   0.000     150      150
#> 239    18     2   407   570      163  949  930   0.000     150      150
#> 240    18     3   657   843      186  944  960   0.000     150      150
#> 241    18     4   887  1133      246  942  961   0.000     150      150
#> 242    18     5  1137  1353      216  970  964   0.000     150      150
#> 243    18     6  1380  1700      320  948  970   0.000     150      150
#> 244    18     7  1730  1900      170 1616  688   0.000     150      150
#> 245    18     8  1903  2530      627 1690  664   0.000     150      150
#> 246    18     9  2563  2753      190 1009  972   0.000     150      150
#> 247    18    10  2756  3010      254  974  987   0.000     150      150
#> 248    18    11  3043  3290      247 1701  656   0.000     150      150
#> 249    18    12  3293  4566     1273 1696  647   0.000     150      150
#> 250    19     1   237   407      170  867  581   0.000     150      150
#> 251    19     2   417   823      406  965  908   0.000     150      150
#> 252    19     3   830  1103      273  984  910   0.000     150      150
#> 253    19     4  1720  2453      733  229  626   0.000     150      150
#> 254    19     5  2456  3016      560  212  644   0.000     150      150
#> 255    19     6  3343  3750      407  201  625   0.000     150      150
#> 256    19     7  3786  4076      290 1008  914   0.000     150      150
#> 257    19     8  4216  4506      290 1606  651   0.000     150      150
#> 258    19     9  4523  4696      173 1144  582   0.000     150      150
#> 259    19    10  4713  5156      443  732  592   0.000     150      150
#> 260    19    11  5173  5936      763 1201  626   0.000     150      150
#> 261    19    12  5953  6556      603  729  616   0.000     150      150
#> 262    19    13  6559  7026      467  710  586   0.000     150      150
#> 263    19    14  7046  7553      507  231  677   0.000     150      150
#> 264    19    15  7566  7833      267  599  625   0.000     150      150
#> 265    19    16  7859  8472      613 1219  633   0.000     150      150
#> 266    20     1     0   163      163 1211  416   0.000     150      150
#> 267    20     2   170   336      166  950  604   0.000     150      150
#> 268    20     3   370   706      336  954  924   0.000     150      150
#> 269    20     4   883  1233      350 1664  661   0.000     150      150
#> 270    20     5  1236  1720      484 1699  654   0.000     150      150
#> 271    20     6  1773  1936      163  403  660   0.000     150      150
#> 272    20     7  2016  2869      853  234  658   0.000     150      150
#> 273    21     1   230   477      247  984  969   0.000     150      150
#> 274    21     2   493   723      230  977  954   0.000     150      150
#> 275    21     3   727  1243      516  974  919   0.000     150      150
#> 276    21     4  1590  1950      360 1234  655   0.000     150      150
#> 277    21     5  1963  2290      327  953  947   0.000     150      150
#> 278    21     6  2327  2560      233  328  708   0.000     150      150
#> 279    21     7  2586  2893      307 1058  827   0.000     150      150
#> 280    21     8  2896  3973     1077  962  963   0.000     150      150
#> 281    21     9  4000  4156      156 1498  673   0.000     150      150
#> 282    21    10  4160  5176     1016 1221  640   0.000     150      150
#> 283    22     1     0   177      177  924  897   0.000     150      150
#> 284    22     2   180   717      537  954  958   0.000     150      150
#> 285    22     3   754  1067      313  925  957   0.000     150      150
#> 286    22     4  1084  1564      480 1178  636   0.000     150      150
#> 287    22     5  1587  2254      667 1687  669   0.000     150      150
#> 288    23     1     0   180      180 1175  672   0.000     150      150
#> 289    23     2   190   370      180  971  906   0.000     150      150
#> 290    23     3   450   697      247  941  953   0.000     150      150
#> 291    23     4   730  1247      517  238  674   0.000     150      150
#> 292    23     5  1277  1430      153  933  894   0.000     150      150
#> 293    23     6  1433  1790      357  961  953   0.000     150      150
#> 294    23     7  1810  2186      376  704  626   0.000     150      150
#> 295    23     8  2207  2456      249  257  655   0.000     150      150
#> 296    23     9  2473  2866      393  732  619   0.000     150      150
#> 297    23    10  2886  3430      544  951  965   0.000     150      150
#> 298    23    11  3463  3696      233  284  651   0.000     150      150
#> 299    23    12  3713  4426      713  713  636   0.000     150      150
#> 300    24     1     0   183      183  971  810   0.000     150      150
#> 301    24     2   187   350      163  961  913   0.000     150      150
#> 302    24     3   357   563      206  931  903   0.000     150      150
#> 303    24     4   593   750      157  267  667   0.000     150      150
#> 304    24     5   753  1250      497  227  651   0.000     150      150
#> 305    25     1   523   793      270  962  949   0.000     150      150
#> 306    25     2   803   966      163  998  958   0.000     150      150
#> 307    25     3  1090  1310      220  976  942   0.000     150      150
#> 308    25     4  1456  1723      267  762  662   0.000     150      150
#> 309    25     5  1736  2766     1030  974  958   0.000     150      150
#> 310    25     6  2769  3536      767  991  971   0.000     150      150
#> 311    25     7  3559  3899      340  732  639   0.000     150      150
#> 312    25     8  3909  4219      310  993  942   0.000     150      150
#> 313    25     9  4223  5046      823  960  967   0.000     150      150
#> 314    25    10  5266  5596      330  716  661   0.000     150      150
#> 315    25    11  5612  6582      970  987  965   0.000     150      150
#> 316    25    12  6626  6782      156  319  606   0.000     150      150
#> 317    25    13  6786  7272      486  223  643   0.000     150      150
#> 318    25    14  7376  7626      250  221  627   0.000     150      150
#> 319    26     1    80   243      163 1236  677   0.000     150      150
#> 320    26     2   270   437      167  576  668   0.000     150      150
#> 321    26     3   453   637      184  962  938   0.000     150      150
#> 322    26     4  1247  1583      336  233  679   0.000     150      150
#> 323    26     5  1600  1790      190  727  618   0.000     150      150
#> 324    26     6  1813  2117      304 1185  632   0.000     150      150
#> 325    26     7  2120  2826      706 1220  655   0.000     150      150
#> 326    26     8  2840  3366      526  945  964   0.000     150      150
#> 327    26     9  3423  4060      637 1226  678   0.000     150      150
#> 328    26    10  4063  4856      793 1222  650   0.000     150      150
#> 329    27     1     0   157      157 1433  605   0.000     150      150
#> 330    27     2   310   474      164  977  921   0.000     150      150
#> 331    27     3   477   844      367  938  947   0.000     150      150
#> 332    27     4   894  1854      960  727  601   0.000     150      150
#> 333    27     5  1860  2037      177 1009  527   0.000     150      150
#> 334    27     6  2040  2437      397  732  576   0.000     150      150
#> 335    28     1    77   267      190  978  762   0.000     150      150
#> 336    28     2   760  1257      497 1631  670   0.000     150      150
#> 337    28     3  1260  1573      313 1642  652   0.000     150      150
#> 338    29     1     0   180      180  844  717   0.000     150      150
#> 339    29     2   207   414      207  973  887   0.000     150      150
#> 340    29     3   507   693      186  961  900   0.000     150      150
#> 341    29     4   727  1677      950  247  660   0.000     150      150
#> 342    29     5  1680  2017      337  219  656   0.000     150      150
#> 343    29     6  2047  2197      150  941  914   0.000     150      150
#> 344    29     7  2200  2550      350  990  973   0.000     150      150
#> 345    30     1     0   157      157  939  720   0.000     150      150
#> 346    30     2   464   917      453 1193  657   0.000     150      150
#> 347    30     3   944  1190      246 1683  653   0.000     150      150
#> 348    30     4  1207  2390     1183 1212  636   0.000     150      150
#> 349    30     5  2407  2757      350  933  929   0.000     150      150
#> 350    30     6  2773  3023      250 1234  642   0.000     150      150
#> 351    30     7  3183  3487      304  718  614   0.000     150      150
#> 352    30     8  3503  4117      614 1221  633   0.000     150      150
#> 353    30     9  4133  4553      420  733  647   0.000     150      150
#> 354    30    10  4557  4747      190  704  634   0.000     150      150
#> 355    31     1    56   247      191  944  716   0.000     150      150
#> 356    31     2   816  1120      304 1004  734   0.000     150      150
#> 357    31     3  1130  1293      163  987  913   0.000     150      150
#> 358    31     4  1330  1576      246 1650  687   0.000     150      150
#> 359    31     5  1580  2403      823 1687  656   0.000     150      150
#> 360    32     1    33   190      157  944  528   0.000     150      150
#> 361    32     2   200   373      173  970  844   0.000     150      150
#> 362    32     3   377   557      180  935  914   0.000     150      150
#> 363    32     4   713   863      150  976  933   0.000     150      150
#> 364    32     5   907  1073      166 1584  634   0.000     150      150
#> 365    32     6  1080  1630      550 1243  600   0.000     150      150
#> 366    33     1     0   150      150  809   17   0.000     150      150
#> 367    33     2   337   704      367  983  950   0.000     150      150
#> 368    33     3   734  1360      626  249  666   0.000     150      150
#> 369    34     1     0   180      180 1189  662   0.000     150      150
#> 370    34     2   480   970      490 1225  617   0.000     150      150
#> 371    34     3   990  1210      220 1644  629   0.000     150      150
#> 372    34     4  1223  1850      627 1215  634   0.000     150      150
#> 373    34     5  1870  2146      276 1665  656   0.000     150      150
#> 374    34     6  2150  2809      659 1692  647   0.000     150      150
#> 375    34     7  2813  4116     1303 1653  581   0.000     150      150
#> 376    34     8  4119  5129     1010 1681  661   0.000     150      150
#> 377    34     9  5163  5339      176  971  689   0.000     150      150
#> 378    34    10  5346  5619      273  739  654   0.000     150      150
#> 379    34    11  5622  6252      630  710  621   0.000     150      150
#> 380    34    12  6272  6819      547 1206  653   0.000     150      150
#> 381    35     1   230   380      150 1116  563   0.239     150      150
#> 382    35     2   517   880      363  974  961   0.000     150      150
#> 383    35     3   963  2160     1197 1685  620   0.000     150      150
#> 384    35     4  2200  2510      310  802  589   0.000     150      150
#> 385    35     5  2513  3203      690  713  609   0.000     150      150
#> 386    35     6  3243  3450      207 1653  654   0.000     150      150
#> 387    35     7  3453  3876      423 1695  629   0.000     150      150
#> 388    35     8  3879  4093      214 1726  480   0.000     150      150
#> 389    35     9  4109  4303      194 1830  196   0.000     150      150
#> 390    35    10  4323  4476      153 1631  598   0.000     150      150
#> 391    35    11  4479  5196      717 1681  632   0.000     150      150
#> 392    35    12  5243  5889      646  749  567   0.000     150      150
#> 393    36     1    46   260      214 1086  652   0.000     150      150
#> 394    36     2   376   536      160  976  879   0.000     150      150
#> 395    36     3   570   846      276 1686  632   0.000     150      150
#> 396    36     4   850  1223      373 1688  625   0.000     150      150
#> 397    37     1     0   160      160  994  621   0.000     150      150
#> 398    37     2   170   336      166  671  647   0.000     150      150
#> 399    37     3   360  1076      716  218  659   0.000     150      150
#> 400    38     1     0   170      170  711  257   0.000     150      150
#> 401    38     2   323   720      397  953  900   0.000     150      150
#> 402    38     3   800  1243      443 1558  628   0.000     150      150
#> 403    38     4  1253  1537      284 1224  662   0.000     150      150
#> 404    38     5  1557  1783      226  940  912   0.000     150      150
#> 405    38     6  1817  2080      263 1674  666   0.000     150      150
#> 406    38     7  2083  2387      304 1691  630   0.000     150      150
#> 407    38     8  2423  2760      337  812  588   0.000     150      150
#> 408    38     9  2773  3523      750 1193  618   0.000     150      150
#> 409    38    10  3553  3893      340 1199  218   0.000     150      150
#> 410    38    11  3916  4403      487  870  579   0.000     150      150
#> 411    38    12  4406  4560      154  736  619   0.000     150      150
#> 412    38    13  4690  4840      150  760  565   0.239     150      150
#> 413    38    14  4843  5450      607  719  587   0.000     150      150
#> 414    38    15  5490  5783      293 1642  625   0.000     150      150
#> 415    38    16  5786  6889     1103 1699  621   0.000     150      150
#> 416    39     1     0   346      346  745  576   0.000     150      150
#> 417    39     2   360   593      233 1177  624   0.000     150      150
#> 418    39     3  1496  1996      500 1207  604   0.000     150      150
#> 419    39     4  2833  2983      150 1140  533   0.239     150      150
#> 420    39     5  2986  3239      253 1230  644   0.000     150      150
#> 421    39     6  3256  3926      670  712  596   0.000     150      150
#> 422    40     1    14   170      156 1118  467   0.000     150      150
#> 423    40     2   174   400      226 1219  636   0.000     150      150
#> 424    40     3   637   994      357 1213  639   0.000     150      150
#> 425    41     1     0   166      166  623  780   0.000     150      150
#> 426    41     2   176   530      354  919  891   0.000     150      150
#> 427    41     3   533   693      160  926  905   0.000     150      150
#> 428    41     4   763  1106      343 1581  672   0.000     150      150
#> 429    41     5  1110  2086      976 1699  653   0.000     150      150
#> 430    42     1     0   173      173 1013  732   0.000     150      150
#> 431    42     2   496   900      404 1201  645   0.000     150      150
#> 432    42     3   920  1276      356  673  615   0.000     150      150
#> 433    42     4  1296  1916      620  282  639   0.000     150      150
#> 434    42     5  1920  2279      359  224  669   0.000     150      150
#> 435    42     6  2310  2476      166  916  896   0.000     150      150
#> 436    42     7  2479  2999      520  959  966   0.000     150      150
#> 437    42     8  3003  3159      156  937  982   0.000     150      150
#> 438    42     9  3219  3466      247  976  958   0.000     150      150
#> 439    42    10  3486  4253      767 1205  649   0.000     150      150
#> 440    42    11  4256  5009      753 1225  638   0.000     150      150
#> 441    43     1     0   177      177  965  551   0.000     150      150
#> 442    43     2   490   787      297  725  638   0.000     150      150
#> 443    44     1     0   314      314  227  636   0.000     150      150
#> 444    45     1     0   180      180  936  809   0.000     150      150
#> 445    45     2   184   370      186  975  945   0.000     150      150
#> 446    45     3   450   730      280  795  633   0.000     150      150
#> 447    45     4   747  1274      527 1254  665   0.000     150      150
#> 448    45     5  1293  1737      444 1629  659   0.000     150      150
#> 449    45     6  1740  2290      550 1697  668   0.000     150      150
#> 450    45     7  2307  2500      193 1231  633   0.000     150      150
#> 451    45     8  2517  4110     1593 1681  644   0.000     150      150
#> 452    46     1     0   174      174 1004  714   0.000     150      150
#> 453    46     2   177   467      290  958  873   0.000     150      150
#> 454    46     3   617  1600      983  245  636   0.000     150      150
#> 455    47     1     0   217      217  859  748   0.000     150      150
#> 456    47     2   223   374      151  961  905   0.000     150      150
#> 457    47     3   380   550      170  958  909   0.000     150      150
#> 458    47     4   570   747      177  712  612   0.000     150      150
#> 459    47     5   750  1183      433  709  612   0.000     150      150
#> 460    47     6  1923  2240      317  702  577   0.000     150      150
#> 461    47     7  2260  2540      280  975  747   0.000     150      150
#> 462    47     8  2730  3066      336  715  599   0.000     150      150
#> 463    47     9  3087  3503      416 1213  602   0.000     150      150
#> 464    47    10  3520  4326      806  718  588   0.000     150      150
#> 465    47    11  4346  4736      390 1228  622   0.000     150      150
#> 466    48     1    57   350      293 1206  651   0.000     150      150
#> 467    49     1     0   163      163 1133  725   0.000     150      150
#> 468    49     2   290   773      483  976  941   0.000     150      150
#> 469    49     3   803  1236      433 1600  696   0.000     150      150
#> 470    49     4  1240  1736      496 1685  672   0.000     150      150
#> 471    50     1     0   213      213 1096  565   0.000     150      150
#> 472    50     2   526   800      274  961  889   0.000     150      150
#> 473    50     3   813  1476      663  717  603   0.000     150      150
#> 474    51     1     0   260      260 1033  729   0.000     150      150
#> 475    51     2   590  1043      453  259  612   0.000     150      150
#> 476    52     1    56   233      177 1024  618   0.000     150      150
#> 477    52     2   240   430      190  999  806   0.000     150      150
#> 478    52     3   436   876      440 1206  605   0.000     150      150
#> 479    53     1     0   220      220  961  863   0.000     150      150
#> 480    53     2   507  1103      596  234  637   0.000     150      150
#> 481    54     1     0   210      210 1159  827   0.000     150      150
#> 482    54     2   286   513      227  944  928   0.000     150      150
#> 483    54     3   546  1396      850 1696  659   0.000     150      150
#> 484    55     1     0   170      170  950  848   0.000     150      150
#> 485    55     2   287   444      157 1185  671   0.000     150      150
#> 486    55     3   447  1177      730 1208  593   0.000     150      150
#> 487    55     4  1190  1574      384  767  614   0.000     150      150
#> 488    55     5  1577  2477      900  714  621   0.000     150      150
#> 489    55     6  2493  2990      497 1224  662   0.000     150      150
#> 490    56     1   560   780      220 1238  667   0.000     150      150
#> 491    56     2   993  1547      554  723  596   0.000     150      150
#> 492    57     1   209   406      197  644  671   0.000     150      150
#> 493    57     2   426   779      353  226  655   0.000     150      150
#> 494    57     3   839  1619      780 1613  629   0.000     150      150
#> 495    58     1   103   253      150  917  614   0.000     150      150
#> 496    58     2   273   573      300  976  922   0.000     150      150
#> 497    58     3   610   886      276  252  636   0.000     150      150
#> 498    58     4   890  1099      209  229  598   0.000     150      150
#> 499    58     5  1103  1509      406  216  645   0.000     150      150
#> 500    59     1     0   190      190 1100  547   0.000     150      150
#> 501    59     2   534  1010      476  752  599   0.000     150      150
#> 502    60     1     0   546      546 1170  613   0.000     150      150
#> 503    61     1    93   263      170 1036  451   0.000     150      150
#> 504    61     2   347   810      463  719   96   0.000     150      150
#> 505    61     3   826  1180      354  360  110   0.000     150      150
#> 506    61     4  1296  1490      194 1161   44   0.000     150      150
#> 507    61     5  1530  1716      186  994 1000   0.000     150      150
#> 508    61     6  1720  1966      246  812 1059   0.000     150      150
#> 509    61     7  2306  2500      194 1151  654   0.000     150      150
#> 510    61     8  2550  2700      150 1668  770   0.000     150      150
#> 511    61     9  2733  2956      223  986  954   0.000     150      150
#> 512    61    10  3259  3499      240  717  622   0.000     150      150
#> 513    61    11  3526  4453      927 1214  626   0.000     150      150
#> 514    61    12  4469  4759      290  756  618   0.000     150      150
#> 515    61    13  4763  5046      283  685  613   0.000     150      150
#> 516    62     1     0   183      183 1174  921   0.000     150      150
#> 517    62     2   190   393      203  979  941   0.000     150      150
#> 518    62     3   406   563      157 1195  704   0.000     150      150
#> 519    62     4   566  1453      887 1212  624   0.000     150      150
#> 520    62     5  1473  1793      320 1639  645   0.000     150      150
#> 521    62     6  1803  1993      190 1224  647   0.000     150      150
#> 522    62     7  2009  2219      210  952  912   0.000     150      150
#> 523    62     8  2233  2553      320 1230  608   0.000     150      150
#> 524    62     9  2593  2969      376  223  648   0.000     150      150
#> 525    62    10  3159  3789      630 1693  648   0.000     150      150
#> 526    62    11  3802  3962      160 1304  610   0.000     150      150
#> 527    62    12  3966  4172      206 1216  643   0.000     150      150
#> 528    63     1   133   306      173 1053  101   0.000     150      150
#> 529    63     2   330   590      260 1228  638   0.000     150      150
#> 530    63     3   896  1046      150  325  683   0.000     150      150
#> 531    63     4  1050  1846      796  222  660   0.000     150      150
#> 532    64     1   140   320      180  922  585   0.000     150      150
#> 533    64     2   327   520      193  976  901   0.000     150      150
#> 534    64     3   547   867      320 1648  621   0.000     150      150
#> 535    64     4   870  1350      480 1690  643   0.000     150      150
#> 536    65     1     0   160      160 1060  851   0.000     150      150
#> 537    65     2   164   397      233  967  931   0.000     150      150
#> 538    65     3   547   740      193 1198  612   0.000     150      150
#> 539    65     4   757  1170      413  763  604   0.000     150      150
#> 540    65     5  1187  1753      566 1224  640   0.000     150      150
#> 541    65     6  1770  2443      673  734  614   0.000     150      150
#> 542    65     7  2447  2640      193  711  608   0.000     150      150
#> 543    65     8  2660  3370      710 1231  655   0.000     150      150
#> 544    65     9  3393  4143      750  695  617   0.000     150      150
#> 545    65    10  4163  4590      427 1235  622   0.000     150      150
#> 546    66     1     0   170      170  965  588   0.000     150      150
#> 547    66     2   403   563      160  699  646   0.000     150      150
#> 548    66     3   586  1126      540  213  646   0.000     150      150
#> 549    67     1     0   150      150 1063  879   0.000     150      150
#> 550    67     2   220   477      257  965  936   0.000     150      150
#> 551    67     3   510   780      270 1585  670   0.000     150      150
#> 552    67     4   783  1213      430 1699  649   0.000     150      150
#> 553    68     1   443   593      150 1144  563   0.239     150      150
#> 554    68     2   739  1509      770  743  591   0.000     150      150
#> 555    68     3  1513  1746      233  704  592   0.000     150      150
#> 556    68     4  1753  2029      276  729  278   0.000     150      150
#> 557    69     1   160   313      153  950  937   0.000     150      150
#> 558    69     2   490   660      170 1199  741   0.000     150      150
#> 559    69     3   663   963      300 1176  661   0.000     150      150
#> 560    69     4  1103  1717      614 1185  625   0.000     150      150
#> 561    70     1   117   280      163  844  581   0.000     150      150
#> 562    70     2   313   507      194  967  905   0.000     150      150
#> 563    70     3   603  1427      824 1633  638   0.000     150      150
#> 564    70     4  1430  1680      250 1698  633   0.000     150      150
#> 565    71     1     0   193      193 1032  685   0.000     150      150
#> 566    71     2   490  1763     1273  721  604   0.000     150      150
#> 567    71     3  1783  2027      244  910  811   0.000     150      150
#> 568    71     4  2030  2320      290 1004  972   0.000     150      150
#> 569    71     5  2323  2727      404  958  938   0.000     150      150
#> 570    71     6  3586  3790      204  964  905   0.000     150      150
#> 571    71     7  4576  4776      200  958  942   0.000     150      150
#> 572    71     8  5610  5979      369  698  598   0.000     150      150
#> 573    71     9  6010  6796      786  214  661   0.000     150      150
#> 574    72     1     0   470      470  200  596   0.000     150      150