Interpolation of missing data (NAs)

Extends the zoo::na.approx and zoo::na.spline functions to include a report which provides the proportion of missing data before and after the interpolation process. This is handy for evaluating the effectiveness of the repair.

Usage

interpolate(data, maxgap = 25, method = "approx", report = FALSE)

Arguments

data

dataframe with columns time, x, y, trial (the standardised raw data form for eyeproc)

maxgap

maximum number of consecutive NAs to fill. Any longer gaps will be left unchanged (see zoo package)

method

"approx" for linear interpolation or "spline" for cubic spline interpolation

report

default is FALSE. If TRUE, then the return value is a list containing the returned data frame and the report.

Value

a dataframe of the same shape of the input data

Examples

interpolate(example_raw_fix, maxgap = 20)
#> # A tibble: 77,669 × 4
#>     time     x     y trial
#>    <dbl> <dbl> <dbl> <dbl>
#>  1     0 1015.  374.     1
#>  2     3 1015.  379.     1
#>  3     7 1013.  376.     1
#>  4    10 1018.  376.     1
#>  5    13 1014.  374.     1
#>  6    17 1020.  376.     1
#>  7    20 1012.  377.     1
#>  8    23 1011.  378.     1
#>  9    27 1009.  372.     1
#> 10    30 1007.  374.     1
#> # ℹ 77,659 more rows
interpolate(example_raw_fix, method = "approx", maxgap = 50, report = TRUE)
#> [[1]]
#> # A tibble: 77,669 × 4
#>     time     x     y trial
#>    <dbl> <dbl> <dbl> <dbl>
#>  1     0 1015.  374.     1
#>  2     3 1015.  379.     1
#>  3     7 1013.  376.     1
#>  4    10 1018.  376.     1
#>  5    13 1014.  374.     1
#>  6    17 1020.  376.     1
#>  7    20 1012.  377.     1
#>  8    23 1011.  378.     1
#>  9    27 1009.  372.     1
#> 10    30 1007.  374.     1
#> # ℹ 77,659 more rows
#> 
#> [[2]]
#> # A tibble: 1 × 2
#>   missing_perc_before missing_perc_after
#>                 <dbl>              <dbl>
#> 1              0.0222             0.0213
#>