Velocity threshold identification of saccades

Use the velocity threshold algorithm from Salvucci & Goldberg (2000) to determine saccadic eye movements. Returns a summary of the saccades found per trial, including start and end coordinates, timing, duration, mean velocity, and peak velocity.

Usage

saccade_VTI(data, threshold = 150, min_dur = 20)

Arguments

data

A dataframe with raw data (pID, time, x, y, trial), the standardised raw data form for eyetools

threshold

velocity threshold (degrees of VA / sec) to be used for identifying saccades

min_dur

minimum duration (ms) expected for saccades. This helps to avoid identification of very short saccades occurring at the boundary of velocity threshold

Value

a data frame giving the saccades found by trial

Details

Analyses data separately for each unique combination of values in pID and trial.

References

Salvucci, D. D., & Goldberg, J. H. (2000). Identifying fixations and saccades in eye-tracking protocols. Proceedings of the Symposium on Eye Tracking Research & Applications - ETRA '00, 71–78.

Examples

data <- combine_eyes(HCL)
saccade_VTI(data)
#>     pID trial sac_n start   end duration  origin_x origin_y terminal_x
#> 1   118     1     1   747   790       43  924.8913 502.8802   572.8710
#> 2   118     1     2  1183  1216       33  550.2522 797.0665   904.1434
#> 3   118     1     3  2193  2243       50 1049.3491 235.9011  1507.8226
#> 4   118     1     4  2690  2750       60 1517.8485 787.2293   972.7982
#> 5   118     1     5  3106  3126       20  951.8168 416.6045  1001.7526
#> 6   118     1     6  3676  3733       57  947.6657 123.4475  1472.9220
#> 7   118     1     7  4223  4279       56 1517.9491 759.8224   973.0053
#> 8   118     1     8  5343  5383       40  928.6512 492.2177   551.4663
#> 9   118     1     9  6792  6826       34  524.8298 829.4756   916.6670
#> 10  118     1    10  7106  7132       26 1008.0858 598.3720  1030.8711
#> 11  118     1    11  7602  7639       37  890.4656 644.2127   524.1420
#> 12  118     1    12  9295  9335       40  573.7701 847.7756   925.2699
#> 13  118     1    13 10435 10489       54  929.5371 135.1185   618.2629
#> 14  118     1    14 10702 10738       36  564.5121 785.2703   945.4900
#> 15  118     1    15 12138 12192       54  977.1132 436.3669  1473.5748
#> 16  118     1    16 12972 13025       53 1538.2589 820.1288   980.1268
#> 17  118     2     1   366   399       33 1009.5214 634.7551  1433.6206
#> 18  118     2     2  1146  1196       50 1488.0968 803.8687  1011.9555
#> 19  118     2     3  1793  1823       30  940.0593 177.6297   936.3069
#> 20  118     2     4  1989  2029       40  899.9465 504.8037   502.1575
#> 21  118     2     5  2626  2679       53  503.8431 787.9458   927.8132
#> 22  118     2     6  3106  3129       23  951.3465 175.7282   971.6648
#> 23  118     2     7  5142  5189       47  990.5532 289.0873  1440.6756
#> 24  118     2     8  5339  5395       56 1439.0196 846.2477   520.0998
#> 25  118     2     9  6235  6285       50  891.8324 167.0562   475.0865
#> 26  118     2    10  6752  6789       37  519.1636 786.9284   939.9464
#> 27  118     2    11  7209  7255       46  989.1180 467.5686  1513.2444
#> 28  118     2    12  7635  7662       27 1513.4448 779.3846  1131.6693
#> 29  118     3     1   933   983       50 1002.7010 194.4611  1485.5097
#> 30  118     3     2  1200  1260       60 1476.1744 841.7218   499.3671
#> 31  118     3     3  1826  1870       44  504.8058 792.4104   914.0560
#> 32  118     3     4  2993  3039       46  996.6983 498.0212  1529.1658
#> 33  118     3     5  3553  3586       33 1453.7949 842.5720  1005.8556
#> 34  118     4     1   160   196       36  981.6367 811.0610  1404.6870
#> 35  118     4     2   540   583       43 1488.3788 767.7919  1037.8731
#> 36  118     4     3  1066  1120       54  932.6748 180.0326   492.1770
#> 37  118     4     4  2389  2443       54 1472.4171 833.8392   637.4518
#> 38  118     4     5  3349  3399       50  508.5161 821.9540   975.5111
#> 39  118     4     6  4156  4216       60  980.6287 176.0979  1388.2071
#> 40  118     4     7  5456  5506       50  509.6492 805.8238   845.4405
#> 41  118     4     8  5839  5859       20  950.2765 202.0675   947.2431
#> 42  118     4     9  6166  6219       53  988.4912 480.0328  1530.9120
#> 43  118     4    10  8009  8072       63 1482.7134 866.4826   505.3686
#> 44  118     4    11  8609  8669       60  595.8946 909.1639  1486.6987
#> 45  118     4    12  8915  8969       54 1493.1766 844.0545   946.8369
#> 46  118     5     1   169   193       24  920.7457 793.6681   665.0876
#> 47  118     5     2   436   469       33  647.7492 768.7432   907.6719
#> 48  118     5     3   766   786       20  907.0568 491.8643   820.8612
#> 49  118     5     4  1059  1113       54  600.8392 824.8404  1462.8866
#> 50  118     5     5  1666  1703       37 1495.7123 778.3008  1060.7272
#> 51  118     5     6  2606  2652       46  964.1227 348.4754  1476.4087
#> 52  118     5     7  2996  3036       40 1529.3674 791.3107  1017.8775
#> 53  118     5     8  3682  3739       57  947.3045 301.1887  1499.1632
#> 54  118     5     9  3926  3979       53 1489.7973 768.8057   957.5302
#> 55  118     6     1   323   343       20  771.3237 683.6370   950.7294
#> 56  118     6     2   577   630       53  964.7887 466.5898  1466.3880
#> 57  118     6     3  1156  1216       60 1484.7630 814.1868   544.1528
#> 58  118     6     4  1526  1580       54  549.6498 756.5507   929.6189
#> 59  118     6     5  2823  2843       20  949.0222 189.3002   933.8608
#> 60  118     6     6  3763  3819       56 1453.7815 824.7820   533.1560
#> 61  118     6     7  4286  4323       37  498.8213 790.8234   925.9516
#> 62  119     1     1   419   439       20  937.4859 596.6131   959.2361
#> 63  119     1     2   963  1009       46  926.6087 448.9321   408.7369
#> 64  119     1     3  1386  1419       33  445.2828 730.4440   774.1377
#> 65  119     1     4  1566  1586       20  821.7980 493.3183   961.6049
#> 66  119     1     5  1859  1886       27  987.8758 332.9815   982.6170
#> 67  119     1     6  2066  2102       36 1008.7884 584.3139  1449.6108
#> 68  119     1     7  3022  3089       67 1475.4953 753.0909   454.0886
#> 69  119     1     8  3292  3326       34  473.3284 717.6178   948.4116
#> 70  119     1     9  4159  4205       46  937.9460 474.0402   483.8080
#> 71  119     1    10  4655  4715       60  505.6226 779.1145  1494.9923
#> 72  119     1    11  5005  5059       54 1433.3890 760.0195   592.4676
#> 73  119     1    12  7301  7371       70  531.4111 724.3928  1471.0353
#> 74  119     1    13  7885  7941       56 1557.4926 706.1470  1031.2273
#> 75  119     1    14  8448  8488       40  940.5212 332.6466   428.2784
#> 76  119     1    15  8848  8901       53  329.9313 650.1681   961.6728
#> 77  119     1    16  9111  9148       37  915.4118 334.5176   506.0964
#> 78  119     1    17  9608  9638       30  491.3408 780.1579   878.9651
#> 79  119     1    18  9984 10024       40  913.6017 468.3063   431.8043
#> 80  119     2     1   160   196       36  930.1504 786.8685   506.3560
#> 81  119     2     2   563   626       63  507.0262 754.6377  1422.9613
#> 82  119     2     3  1086  1123       37 1536.6883 737.5674  1171.5084
#> 83  119     2     4  2969  2999       30  958.4742 287.2842   936.6324
#> 84  119     2     5  3216  3249       33  904.4740 653.2866   534.9389
#> 85  119     2     6  3563  3602       39  529.6488 744.3024   943.8313
#> 86  119     2     7  4002  4026       24  935.2114 349.1000   813.9743
#> 87  119     2     8  4216  4256       40  741.5883 472.8151   527.5179
#> 88  119     2     9  4462  4522       60  565.8551 782.8804  1449.5204
#> 89  119     3     1   187   220       33  929.6976 731.8546   491.3072
#> 90  119     3     2   663   727       64  475.6411 761.7902  1465.4223
#> 91  119     3     3  1167  1197       30 1388.1631 659.8787  1009.4100
#> 92  119     3     4  1647  1670       23  956.9901 379.3481   996.2286
#> 93  119     3     5  2226  2280       54 1083.8245 440.7186  1366.5362
#> 94  119     3     6  3113  3180       67 1499.9854 787.7492   475.7855
#> 95  119     3     7  3929  3993       64  448.9802 745.1622  1460.1439
#> 96  119     3     8  4989  5069       80 1359.4973 693.7878   484.6122
#> 97  119     3     9  5093  5119       26  446.2522 690.6419   394.5154
#> 98  119     3    10  5692  5746       54  451.1674 677.6303  1024.6236
#> 99  119     3    11  7135  7162       27  933.4311 313.1571   735.9905
#> 100 119     3    12  7365  7405       40  858.5582 358.8731   506.7146
#> 101 119     4     1   237   270       33  926.5049 725.0489   514.1258
#> 102 119     4     2  1390  1440       50  553.2898 758.5523  1410.9398
#> 103 119     4     3  1853  1896       43 1400.3527 761.2807   959.3316
#> 104 119     4     4  2993  3026       33  410.2407 759.6056   834.9296
#> 105 119     4     5  3256  3296       40  878.8572 486.6256   511.0056
#> 106 119     4     6  4143  4199       56  545.5005 791.4970  1429.2845
#> 107 119     4     7  5416  5479       63 1449.6049 722.3000   503.8289
#> 108 119     4     8  6066  6119       53  538.4092 777.9468  1352.3635
#> 109 119     4     9  6625  6682       57 1457.5757 735.3299   509.4560
#> 110 119     4    10  7355  7405       50  455.0061 766.4567  1029.1588
#> 111 119     4    11  7749  7769       20  962.0985 371.5052   975.0627
#> 112 119     4    12  8605  8625       20  961.2967 482.5752   946.2679
#> 113 119     5     1   214   250       36  929.7282 740.4175   495.4903
#> 114 119     5     2  1100  1160       60  562.2012 777.7919  1469.1228
#> 115 119     5     3  2193  2227       34 1623.7868 783.7612  1626.5978
#> 116 119     5     4  2397  2450       53 1624.6249 782.5055  1623.9142
#> 117 119     5     5  2470  2523       53 1626.5158 776.7604  1624.3390
#> 118 119     5     6  2720  2796       76 1575.6467 801.0315   580.4435
#> 119 119     5     7  2890  2916       26  434.0244 734.8002   448.7737
#> 120 119     5     8  3386  3456       70  487.2023 770.7846  1464.8231
#> 121 119     5     9  4499  4569       70 1337.8185 786.2550   518.6549
#> 122 119     5    10  5519  5596       77  559.5449 750.5246  1466.2904
#> 123 119     5    11  6606  6686       80 1538.5927 769.7452   532.0211
#> 124 119     5    12  7395  7429       34  561.1405 727.3658   929.0725
#> 125 119     5    13  8215  8235       20  959.0783 363.6570  1002.1730
#> 126 119     5    14  9009  9055       46  966.1764 181.1402   490.9862
#> 127 119     5    15  9565  9592       27  514.3670 682.0681   863.4544
#> 128 119     5    16 10575 10615       40  863.9232 536.9596   477.5950
#> 129 119     5    17 11755 11798       43  431.9173 762.7475   895.9795
#> 130 119     5    18 12374 12401       27  928.1233 339.4489   708.4844
#> 131 119     6     1   590   626       36  893.5909 421.1214   445.4796
#> 132 119     6     2  3053  3106       53  456.9632 729.2448  1304.1947
#> 133 119     6     3  4232  4262       30 1372.3698 735.7928  1024.0037
#> 134 119     6     4  6778  6818       40 1008.7615 446.5270  1449.8786
#> 135 119     6     5  6832  6852       20 1556.7990 621.5306  1709.8604
#> 136 119     6     6  6938  6965       27 1635.7978 766.4061  1625.8207
#> 137 119     6     7  7558  7588       30 1524.0060 686.7242  1142.0919
#> 138 119     6     8  9365  9415       50 1000.8005 293.0768  1397.7841
#> 139 119     6     9 10281 10344       63 1419.5092 783.3005   522.3444
#> 140 119     6    10 11021 11084       63  495.6926 799.3568  1323.7945
#> 141 119     6    11 11581 11644       63 1488.0635 757.2932   501.3444
#> 142 119     6    12 11994 12021       27  469.7133 707.4822   724.9794
#> 143 119     6    13 12034 12064       30  775.1127 552.5540   998.6819
#>     terminal_y mean_velocity peak_velocity
#> 1    808.04272      283.5040      412.5845
#> 2    572.85360      330.0384      473.0425
#> 3    815.95874      403.2042      651.9966
#> 4    362.66265      315.1539      489.6344
#> 5    182.75663      316.1650      434.0858
#> 6    773.01410      389.7924      705.5829
#> 7    344.40579      328.1166      550.8807
#> 8    811.74409      318.4488      469.8318
#> 9    655.82093      331.7420      452.1227
#> 10   647.37457      443.3985      847.4330
#> 11   812.66139      288.3583      401.9520
#> 12   507.11854      322.8852      502.7726
#> 13   687.51196      312.7039      574.2841
#> 14   500.89552      333.6369      503.8933
#> 15   820.45975      307.2883      451.3544
#> 16   421.15891      346.3010      579.4354
#> 17   800.74046      352.2269      556.1337
#> 18   458.93763      332.4957      519.3703
#> 19   499.25743      284.7993      429.6438
#> 20   806.82680      331.5106      500.1617
#> 21   253.97420      334.2068      531.0173
#> 22   465.01123      310.5326      558.4837
#> 23   817.47684      392.6976      715.2217
#> 24   853.99013      427.6617      673.6405
#> 25   749.27484      378.3523      608.5170
#> 26   510.38435      351.4051      499.7125
#> 27   806.21239      348.2006      517.1956
#> 28   548.60528      420.9308      653.0899
#> 29   811.01401      408.7409      734.1952
#> 30   866.36876      418.6330      696.7679
#> 31   487.25831      306.0391      519.2672
#> 32   836.79907      362.6508      544.8551
#> 33   514.52377      512.3510      980.4846
#> 34   807.59300      299.5884      564.8238
#> 35   407.89316      355.1235      573.0521
#> 36   849.41309      388.0088      711.3487
#> 37   817.06385      409.4915      586.4790
#> 38   465.23668      304.9752      470.5284
#> 39   770.51439      322.6836      665.8114
#> 40   270.38513      328.6062      522.0221
#> 41   452.90632      325.6190      467.3367
#> 42   892.25707      336.0883      544.9523
#> 43   889.74747      406.8992      619.2330
#> 44   885.09883      383.8644      521.0181
#> 45   416.29324      339.6442      541.8968
#> 46   815.40923      277.2199      359.5104
#> 47   460.19167      322.8844      570.6843
#> 48   647.49100      251.7870      277.0848
#> 49   881.41554      417.7221      575.1774
#> 50   485.14086      377.9888      603.7192
#> 51   741.68104      362.0846      606.5546
#> 52   393.74462      419.1947      584.4507
#> 53   815.85457      347.7312      706.5983
#> 54   181.31785      392.0532      522.5838
#> 55   494.17258      337.6080      456.3115
#> 56   819.45070      309.3438      471.1991
#> 57   813.94843      402.1739      590.0195
#> 58   311.93170      294.9704      437.3743
#> 59   411.30941      283.6373      367.7784
#> 60   830.62670      424.8246      615.6619
#> 61   578.07589      332.7554      496.5516
#> 62   385.89048      271.9732      358.4825
#> 63   736.63587      338.0463      452.5216
#> 64   522.73404      309.1227      427.8199
#> 65   332.33636      266.9573      444.8036
#> 66   557.67814      233.1243      341.8791
#> 67   731.04919      333.5676      447.3443
#> 68   723.96569      399.4037      511.9482
#> 69   530.82351      399.0647      613.1668
#> 70   726.20847      300.6745      390.4783
#> 71   785.48033      432.6330      550.8206
#> 72   525.59939      493.7518      774.3539
#> 73   765.71968      355.9139      477.1906
#> 74   277.11461      326.7673      455.2461
#> 75   636.55022      391.6198      520.0229
#> 76   303.23776      356.3590      648.3147
#> 77   685.16739      386.1747      471.2884
#> 78   529.08806      407.7882      575.6108
#> 79   724.13881      366.2645      480.4402
#> 80   762.26611      310.1253      444.8956
#> 81   771.04969      373.1909      464.0138
#> 82   521.52158      298.5782      413.8839
#> 83   584.14036      285.9511      424.2707
#> 84   773.08285      304.8708      520.0391
#> 85   234.80312      419.1758      618.5850
#> 86   447.56302      192.7987      236.2379
#> 87   786.49801      265.8744      337.9217
#> 88   750.54686      391.1909      529.0502
#> 89   758.72776      348.0018      448.1696
#> 90   706.53287      410.5039      574.1005
#> 91   416.62787      405.3553      510.3707
#> 92   120.06360      295.9172      541.4810
#> 93   692.01373      214.5089      305.3467
#> 94   749.79324      403.1262      523.6552
#> 95   739.73714      424.7965      560.3353
#> 96   624.15468      389.4591     1161.9753
#> 97   683.99888      483.3973     1146.8924
#> 98   445.97947      321.0485      473.3812
#> 99   401.45857      224.3269      342.5730
#> 100  760.78629      348.9148      466.7990
#> 101  746.55221      322.2882      442.0032
#> 102  760.38561      453.1791      595.6031
#> 103  585.73589      333.0678      675.7288
#> 104  477.50542      391.8615      507.1958
#> 105  735.42724      293.9378      445.6026
#> 106  725.27627      408.2254      560.2054
#> 107  750.76249      387.8402      487.5617
#> 108  565.98955      460.6720      821.0242
#> 109  768.40618      444.2416      549.3145
#> 110  449.86136      343.2085      455.5196
#> 111   87.88856      353.4739      533.5421
#> 112  279.86526      316.7568      568.0781
#> 113  735.73645      311.7389      429.0032
#> 114  745.15599      399.2416      496.8057
#> 115  781.91854      220.1227      247.4348
#> 116  780.19254      196.1123      229.7452
#> 117  796.44055      201.9337      317.7640
#> 118  621.57478      401.7760     1082.7339
#> 119  794.40469      558.0969     1265.4959
#> 120  733.85336      368.1586      521.4750
#> 121  771.84247      304.1642      399.6805
#> 122  769.89695      320.2727      443.9788
#> 123  627.22944      425.9321      659.6841
#> 124  426.38805      391.0488      519.2551
#> 125  150.70444      285.3563      389.7502
#> 126  573.53000      354.8389      531.0020
#> 127  402.05052      429.9127      554.2282
#> 128  694.73338      286.3361      401.5039
#> 129  340.18672      381.8093      617.3080
#> 130  422.20742      336.2447      472.3969
#> 131  719.21666      391.3893      477.2416
#> 132  747.38857      433.2943      511.7817
#> 133  514.08373      384.9760      475.9437
#> 134  602.33134      304.8740      357.7085
#> 135  546.70075      529.5197     1399.8635
#> 136  792.84668      239.9725      327.4583
#> 137  403.44032      417.6347      516.2530
#> 138  657.71066      296.4511      420.2386
#> 139  756.58455      368.9933      481.3838
#> 140  779.42050      343.6269      513.4852
#> 141  729.33815      409.3816      518.1208
#> 142  584.69650      275.5924      410.6422
#> 143  409.76345      256.7255      411.7965