Sequence analysis of area of interest entries

Analyses the sequence of entries into defined AOI regions across trials. Works with fixation data or raw data as the input.

Usage

AOI_seq(data, AOIs, AOI_names = NULL, sample_rate = NULL)

Arguments

data

A dataframe with fixation data (from fix_dispersion) or raw data.

AOIs

A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).

AOI_names

An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.

sample_rate

Optional sample rate of the eye-tracker (Hz) for use with raw_data. If not supplied, the sample rate will be estimated from the time column and the number of samples.

Value

a dataframe containing the sequence of entries into AOIs on each trial

Examples

fix_d <- fix_dispersion(eyetools::example_raw_WM)
AOI_seq(fix_d, eyetools::AOIs_WM)
#>     trial AOI_entry_seq
#> 1       1   3;4;3;2;4;2
#> 2       2     4;3;1;2;4
#> 3       3     2;4;3;1;3
#> 4       4   1;2;4;3;1;4
#> 5       5     1;2;1;2;4
#> 6       6       2;4;2;1
#> 7       7 2;4;2;3;4;2;1
#> 8       8         2;4;1
#> 9       9     3;1;3;1;2
#> 10     10         2;1;3
#> 11     11     4;3;1;3;4
#> 12     12     2;4;2;1;3
#> 13     13       3;2;1;3
#> 14     14     1;3;1;3;1
#> 15     15       4;2;1;3
#> 16     16       1;3;4;2
#> 17     17   2;1;3;4;1;2
#> 18     18         1;3;4
#> 19     19         4;2;1
#> 20     20       2;4;1;3
#> 21     21       4;2;4;3
#> 22     22     4;2;4;2;3
#> 23     23     3;1;3;4;2
#> 24     24       2;4;2;1
#> 25     25         1;3;2
#> 26     26         2;4;2
#> 27     27   2;4;2;4;2;1
#> 28     28           2;4
#> 29     29       4;2;4;2
#> 30     30   2;4;2;4;1;4
#> 31     31   4;3;2;3;1;4
#> 32     32       4;2;4;2
#> 33     33       2;4;1;3
#> 34     34       3;1;2;4
#> 35     35             2
#> 36     36         2;4;3
#> 37     37     2;4;1;3;2
#> 38     38   2;4;2;4;1;3
#> 39     39   3;2;1;2;4;3
#> 40     40   4;2;1;3;4;2
#> 41     41         1;3;4
#> 42     42         4;2;1
#> 43     43     3;2;1;3;1
#> 44     44   2;4;2;4;2;4
#> 45     45     2;4;2;4;1
#> 46     46     2;4;2;4;1
#> 47     47   2;4;2;1;2;3
#> 48     48       2;4;2;4
#> 49     49     4;2;4;2;1
#> 50     50     4;3;1;2;4
#> 51     51         1;2;1
#> 52     52     4;2;4;2;3
#> 53     53         1;2;4
#> 54     54       3;1;2;4
#> 55     55     4;2;4;2;4
#> 56     56       2;4;1;3
#> 57     57       1;3;4;2
#> 58     58     4;2;4;2;1
#> 59     59       4;2;1;3
#> 60     60     2;4;1;3;2
#> 61     61     2;4;2;1;3
#> 62     62             1
#> 63     63   4;2;4;2;1;3
#> 64     64     2;4;2;1;3
#> 65     65   1;3;1;3;4;2
#> 66     66   1;2;1;3;4;2
#> 67     67         2;4;1
#> 68     68     2;4;2;4;1
#> 69     69       4;3;1;2
#> 70     70       1;2;4;3
#> 71     71     4;2;4;2;1
#> 72     72       4;2;1;4
#> 73     73   1;3;4;2;4;1
#> 74     74         3;1;2
#> 75     75       2;4;3;2
#> 76     76     4;2;1;3;4
#> 77     77       3;1;2;4
#> 78     78         4;2;1
#> 79     79     1;2;4;3;1
#> 80     80       1;3;4;2
#> 81     81     2;4;2;4;1
#> 82     82         1;2;4
#> 83     83       3;2;4;1
#> 84     84         2;4;1
#> 85     85         2;4;1
#> 86     86           3;1
#> 87     87       4;2;4;2
#> 88     88     4;2;4;2;1
#> 89     89         3;1;4
#> 90     90     2;4;2;4;1
#> 91     91           3;1
#> 92     92     2;4;2;1;3
#> 93     93     1;4;3;4;2
#> 94     94       2;4;1;3
#> 95     95         2;4;1
#> 96     96         3;1;4
#> 97     97         3;1;3
#> 98     98       1;3;4;2
#> 99     99       4;3;2;3
#> 100   100             1