Time analysis of area of interest entries

Analyses total time on defined AOI regions across trials. Works with fixation and raw data as the input (must use one or the other, not both).

Usage

AOI_time(
  fix_data = NULL,
  raw_data = NULL,
  AOIs,
  AOI_names = NULL,
  sample_rate = NULL
)

Arguments

fix_data

A dataframe with fixation data (from fix_dispersion)

raw_data

A dataframe with raw data

AOIs

A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).

AOI_names

An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.

sample_rate

Optional sample rate of the eye-tracker (Hz) for use with raw_data. If not supplied, the sample rate will be estimated from the time column and the number of samples.

Value

a dataframe containing the time on the passed AOIs for each trial

Examples

fix_d <- fix_dispersion(eyetools::example_raw_WM)
AOI_time(fix_data = fix_d, AOIs = eyetools::AOIs_WM)
#>     trial AOI_1 AOI_2 AOI_3 AOI_4
#> 1       1     0   617   583   383
#> 2       2   384   334   442   583
#> 3       3   350   325   475   366
#> 4       4   317   250   250   633
#> 5       5   782   600     0   159
#> 6       6   291   734     0   816
#> 7       7   225   717   175   350
#> 8       8   275  1009     0   375
#> 9       9   493   241   775     0
#> 10     10   433   533   250     0
#> 11     11   441     0   708   433
#> 12     12   258   750   225   358
#> 13     13   451   300   808     0
#> 14     14  1343     0   384     0
#> 15     15   284   416   167   267
#> 16     16   525   425   300   150
#> 17     17   734   384   408   159
#> 18     18   650     0   425   300
#> 19     19   784   300     0   509
#> 20     20   283   642   250   383
#> 21     21     0   466   342   825
#> 22     22     0   567   400   783
#> 23     23   283   192   483   233
#> 24     24   366   868     0   300
#> 25     25  1033   183   491     0
#> 26     26     0  1316     0   542
#> 27     27   532   675     0   576
#> 28     28     0  1242     0   542
#> 29     29     0   425     0  1042
#> 30     30   275   567     0   683
#> 31     31   233   309   726   459
#> 32     32     0  1082     0   767
#> 33     33   167   617   292   525
#> 34     34   600   400   341   250
#> 35     35     0  1575     0     0
#> 36     36     0   733   425   468
#> 37     37   150   982   216   333
#> 38     38   375   659   267   467
#> 39     39   267   542   716   150
#> 40     40   225   859   242   483
#> 41     41   341     0   592   584
#> 42     42   408   692     0   550
#> 43     43   350   283   859     0
#> 44     44     0   817     0   691
#> 45     45   558   717     0   474
#> 46     46   259   726     0   766
#> 47     47   300   976   267   241
#> 48     48     0   542     0   458
#> 49     49   366   666     0   616
#> 50     50   317   342   233   784
#> 51     51  1334   534     0     0
#> 52     52     0   425   208   751
#> 53     53   666   242     0   667
#> 54     54   333   341   592   258
#> 55     55     0   625     0   884
#> 56     56   416   941   167   175
#> 57     57  1042   358   258   233
#> 58     58   167   749     0   717
#> 59     59   233   876   267   467
#> 60     60   208   933   208   334
#> 61     61   333   732   175   576
#> 62     62  1909     0     0     0
#> 63     63   234   441   474   650
#> 64     64   384   859   300   225
#> 65     65   725   275   509   208
#> 66     66   682   358   375   258
#> 67     67   542   640     0   525
#> 68     68   317   625     0   566
#> 69     69   300   658   641   258
#> 70     70   709   351   375   417
#> 71     71   225   350     0  1149
#> 72     72   208   741     0   800
#> 73     73   593   208   275   450
#> 74     74   767   350   608     0
#> 75     75     0  1276   217   309
#> 76     76   208   392   158   574
#> 77     77   666   275   234   409
#> 78     78   217   817     0   666
#> 79     79   374   350   275   275
#> 80     80   726   275   217   658
#> 81     81   400   925     0   416
#> 82     82   317   158     0   633
#> 83     83   459   259   334   625
#> 84     84   375  1090     0   375
#> 85     85   250   941     0   692
#> 86     86   375     0  1001     0
#> 87     87     0   500     0   992
#> 88     88   183   625     0   800
#> 89     89   691     0   750   317
#> 90     90   184   659     0   958
#> 91     91   200     0  1733     0
#> 92     92   167   533   225   858
#> 93     93   233   225   534   475
#> 94     94   333   833   175   317
#> 95     95   333   858     0   516
#> 96     96   192     0  1058   234
#> 97     97   350     0  1257     0
#> 98     98   616   325   275   184
#> 99     99     0   476   733   266
#> 100   100  1999     0     0     0
AOI_time(raw_data = eyetools::example_raw_WM, AOIs = eyetools::AOIs_WM, sample_rate = 120)
#>     trial AOI_1 AOI_2 AOI_3 AOI_4
#> 1       1    17   667   625   425
#> 2       2   408   375   542   642
#> 3       3   425   358   600   392
#> 4       4   400   425   275   683
#> 5       5   817   742     0   333
#> 6       6   308   800     0   850
#> 7       7   283   808   283   400
#> 8       8   300  1100     0   550
#> 9       9   550   425   908     0
#> 10     10   517   775   283   358
#> 11     11   483   117   742   617
#> 12     12   275   833   233   358
#> 13     13   483   325   842   150
#> 14     14  1542     0   425     0
#> 15     15   308   467   175   658
#> 16     16   567   492   325   417
#> 17     17   800   408   442   200
#> 18     18   675   450   450   392
#> 19     19   817   342   275   525
#> 20     20   400   733   258   550
#> 21     21     0   500   358  1050
#> 22     22    25   642   450   817
#> 23     23   325   358   542   417
#> 24     24   392  1075   133   333
#> 25     25  1092   208   500     0
#> 26     26    33  1383     0   542
#> 27     27   558   733     0   617
#> 28     28     0  1292     0   558
#> 29     29     0   475    92  1225
#> 30     30   308   617   150   792
#> 31     31   258   350   800   517
#> 32     32     0  1158     0   817
#> 33     33   417   642   325   558
#> 34     34   642   417   375   275
#> 35     35     0  1733     0     0
#> 36     36   233   767   425   517
#> 37     37   167  1067   217   467
#> 38     38   400   708   308   500
#> 39     39   292   600   842   192
#> 40     40   250   917   242   533
#> 41     41   375    67   608   617
#> 42     42   450   750     0   742
#> 43     43   400   400  1042    25
#> 44     44     0  1025   133   750
#> 45     45   592   775    42   550
#> 46     46   283   800     0   808
#> 47     47   325  1017   292   267
#> 48     48     0   733   133   533
#> 49     49   408   758   125   633
#> 50     50   392   383   275   833
#> 51     51  1400   575     0     0
#> 52     52   108   542   225   817
#> 53     53   692   417   158   708
#> 54     54   375   383   625   292
#> 55     55     0   767     0  1000
#> 56     56   425  1042   175   242
#> 57     57  1050   358   283   258
#> 58     58   233   825     0   900
#> 59     59   242   942   275   492
#> 60     60   242  1092   233   350
#> 61     61   367   792   200   608
#> 62     62  2017     0     0     0
#> 63     63   258   500   483   692
#> 64     64   425   908   317   292
#> 65     65   842   283   558   250
#> 66     66   717   383   517   283
#> 67     67   558   792    33   575
#> 68     68   467   675     0   800
#> 69     69   333   700   658   275
#> 70     70   725   392   392   450
#> 71     71   242   417    75  1167
#> 72     72   217   792     0   883
#> 73     73   650   250   308   483
#> 74     74   800   358   625   158
#> 75     75     0  1333   233   375
#> 76     76   225   475   183   625
#> 77     77   700   283   250   417
#> 78     78   375   858     0   708
#> 79     79   517   508   300   308
#> 80     80   750   300   250   692
#> 81     81   500   992     0   475
#> 82     82   367   225   117   667
#> 83     83   642   292   333   642
#> 84     84   400   992     0   442
#> 85     85   275   975     0   733
#> 86     86   400     0   892    75
#> 87     87     0   542     0  1108
#> 88     88   342   708    58   842
#> 89     89   733     0   800   333
#> 90     90   242   692     0  1000
#> 91     91   200     0  1783     0
#> 92     92   183   575   250   908
#> 93     93   275   267   575   575
#> 94     94   358   858   192   567
#> 95     95   358   908     0   558
#> 96     96   233   292  1150   267
#> 97     97   392     0  1350     0
#> 98     98   817   358   308   208
#> 99     99     0   525   800   283
#> 100   100  2017     0     0     0

#' @importFrom dplyr between