Fixation detection using a velocity threshold identification method

Determine fixations by assessing the velocity of eye-movements, using a method that is similar to that proposed by Salvucci & Goldberg (1996). Applies the algorithm used in VTI_saccade and removes the identified saccades before assessing whether separated fixations are outside of the dispersion tolerance. If they are outside of this tolerance, the fixation is treated as a new fixation regardless of the length of saccade separating them. Compared to fixation_dispersion(), fixation_VTI() is more conservative in determining a fixation as smaller saccades are discounted and the resulting data is treated as a continued fixation (assuming it is within the pixel tolerance set by disp_tol). Returns a summary of the fixations found per trial, including start and end coordinates, timing, duration, mean velocity, and peak velocity.

Usage

fixation_VTI(
  data,
  sample_rate = NULL,
  threshold = 100,
  min_dur = 150,
  min_dur_sac = 20,
  disp_tol = 100,
  run_interp = TRUE,
  smooth = FALSE,
  progress = TRUE,
  participant_ID = "participant_ID"
)

Arguments

data

A dataframe with raw data (time, x, y, trial) for one participant

sample_rate

sample rate of the eye-tracker. If default of NULL, then it will be computed from the timestamp data and the number of samples

threshold

velocity threshold (degrees of VA / sec) to be used for identifying saccades.

min_dur

Minimum duration (in milliseconds) of period over which fixations are assessed

min_dur_sac

Minimum duration (in milliseconds) for saccades to be determined

disp_tol

Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period

run_interp

include a call to eyetools::interpolate on each trial.

smooth

include a call to eyetools::smoother on each trial

progress

Display a progress bar

participant_ID

the variable that determines the participant identifier. If no column present, assumes a single participant

Value

a dataframe containing each detected fixation by trial, with mean x/y position in pixel, start and end times, and duration.

Examples

# single trials:
fixation_VTI(example_raw_WM[example_raw_WM$trial == 15,])
#>   trialNumber fix_n start  end duration         x        y min_dur disp_tol
#> 1          15     1   266  500      234 1310.5204 204.2080     150      100
#> 2          15     2   541  758      217  989.9692 577.2593     150      100
#> 3          15     3   808 1225      417  550.2882 161.0689     150      100
#> 4          15     4  1275 1466      191 1164.3131 125.7277     150      100
#> 5          15     5  1525 1808      283  648.5091 882.2266     150      100
#> 6          15     6  1850 2008      158 1281.9817 879.6660     150      100

# multiple trials:

fixation_VTI(example_raw_WM[example_raw_WM$trial %in% c(16),])
#>   trialNumber fix_n start  end duration         x        y min_dur disp_tol
#> 1          16     1     0  175      175  627.5234 622.3245     150      100
#> 2          16     2   208  725      517  630.6710 875.2348     150      100
#> 3          16     3   775 1075      300 1310.4355 896.5777     150      100
#> 4          16     4  1108 1500      392 1335.3174 146.6219     150      100
#> 5          16     5  1567 2008      441  571.4950 203.7527     150      100

#' @references Salvucci, D. D., & Goldberg, J. H. (2000). Identifying fixations and saccades in eye-tracking protocols. Proceedings of the Symposium on Eye Tracking Research & Applications - ETRA ’00, 71–78.