Fixation detection using a velocity threshold identification method

Determine fixations by assessing the velocity of eye-movements, using a method that is similar to that proposed by Salvucci & Goldberg (1996). Applies the algorithm used in VTI_saccade and removes the identified saccades before assessing whether separated fixations are outside of the dispersion tolerance. If they are outside of this tolerance, the fixation is treated as a new fixation regardless of the length of saccade separating them. Compared to fixation_dispersion(), fixation_VTI() is more conservative in determining a fixation as smaller saccades are discounted and the resulting data is treated as a continued fixation (assuming it is within the pixel tolerance set by disp_tol). Returns a summary of the fixations found per trial, including start and end coordinates, timing, duration, mean velocity, and peak velocity.

Usage

fixation_VTI(
  data,
  sample_rate = NULL,
  threshold = 100,
  min_dur = 150,
  min_dur_sac = 20,
  disp_tol = 100,
  run_interp = TRUE,
  smooth = FALSE,
  progress = TRUE
)

Arguments

data

A dataframe with raw data (time, x, y, trial) for one participant

sample_rate

sample rate of the eye-tracker. If default of NULL, then it will be computed from the timestamp data and the number of samples

threshold

velocity threshold (degrees of VA / sec) to be used for identifying saccades.

min_dur

Minimum duration (in milliseconds) of period over which fixations are assessed

min_dur_sac

Minimum duration (in milliseconds) for saccades to be determined

disp_tol

Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period

run_interp

include a call to eyetools::interpolate on each trial.

smooth

include a call to eyetools::smoother on each trial

progress

Display a progress bar

Value

a dataframe containing each detected fixation by trial, with mean x/y position in pixel, start and end times, and duration.

Examples

# single trials:
fixation_VTI(example_raw_WM[example_raw_WM$trial == 15,])
#>   trial fix_n start  end duration         x        y min_dur disp_tol
#> 1    15     1   266  500      234 1310.5204 204.2080     150      100
#> 2    15     2   541  758      217  989.9692 577.2593     150      100
#> 3    15     3   808 1225      417  550.2882 161.0689     150      100
#> 4    15     4  1275 1466      191 1164.3131 125.7277     150      100
#> 5    15     5  1525 1808      283  648.5091 882.2266     150      100
#> 6    15     6  1850 2008      158 1281.9817 879.6660     150      100

# multiple trials:
data <- rbind(example_raw_WM[example_raw_WM$trial %in% c(3:10),])

fixation_VTI(data)
#>    trial fix_n start  end duration         x        y min_dur disp_tol
#> 1      3     1     0  166      166  955.4684 430.8143     150      100
#> 2      3     2   200  516      316  575.5794 219.3148     150      100
#> 3      3     3   558  916      358 1350.6525 172.6407     150      100
#> 4      3     4   958 1166      208 1172.2551 882.3413     150      100
#> 5      3     5  1191 1583      392  617.2695 905.9970     150      100
#> 6      3     6  1633 2008      375 1236.2489 835.5535     150      100
#> 7      4     1     0  166      166  831.8167 636.9694     150      100
#> 8      4     2   191  358      167  616.5438 903.4325     150      100
#> 9      4     3   550  791      241  579.8265 237.5259     150      100
#> 10     4     4   833 1241      408 1250.9701 148.6456     150      100
#> 11     4     5  1291 1541      250 1252.0957 855.6089     150      100
#> 12     4     6  1583 1733      150  628.5860 845.0303     150      100
#> 13     4     7  1799 2008      209 1302.5259 366.7057     150      100
#> 14     5     1     0  416      416  623.1304 875.9831     150      100
#> 15     5     2   466  741      275  474.9025 210.7790     150      100
#> 16     5     3   908 1266      358  594.8604 739.2825     150      100
#> 17     5     4  1316 1616      300  509.8356 185.7359     150      100
#> 18     5     5  1825 1975      150 1334.0981 181.2156     150      100
#> 19     6     1     0  167      167  638.1851 275.9460     150      100
#> 20     6     2   200  550      350  546.8152 189.1050     150      100
#> 21     6     3   592  775      183 1076.5818 152.0481     150      100
#> 22     6     4   800 1433      633 1332.0275 166.9014     150      100
#> 23     6     5  1467 1675      208  593.2186 167.6117     150      100
#> 24     6     6  1717 2008      291  602.5118 793.9582     150      100
#> 25     7     1     0  283      283  563.5335 202.3064     150      100
#> 26     7     2   325  508      183 1390.6734 239.0201     150      100
#> 27     7     3   700  916      216  654.6657 204.9952     150      100
#> 28     7     4   958 1133      175 1274.2894 834.4385     150      100
#> 29     7     5  1166 1341      175 1240.0263 137.7849     150      100
#> 30     7     6  1375 1608      233  592.8054 180.4172     150      100
#> 31     7     7  1641 1908      267  613.2523 872.1690     150      100
#> 32     8     1    50  550      500  569.5150 204.8009     150      100
#> 33     8     2   775 1266      491  539.6886 174.7308     150      100
#> 34     8     3  1308 1683      375 1311.2686 185.4978     150      100
#> 35     8     4  1741 2008      267  664.9630 882.1257     150      100
#> 36     9     1     0  717      717 1295.2336 893.3710     150      100
#> 37     9     2   767 1042      275  619.0256 903.5558     150      100
#> 38     9     3  1083 1242      159 1274.1981 890.0653     150      100
#> 39     9     4  1283 1492      209  633.3485 898.4240     150      100
#> 40     9     5  1542 1767      225  497.2668 200.4549     150      100
#> 41     9     6  1792 2000      208  638.4807 193.4646     150      100
#> 42    10     1     0  175      175  702.5992 294.8774     150      100
#> 43    10     2   208  466      258  548.0793 204.1055     150      100
#> 44    10     3   550  841      291 1278.3503 153.2124     150      100
#> 45    10     4   891 1116      225  677.2301 152.0584     150      100
#> 46    10     5  1183 1624      441  644.2483 855.3782     150      100
#> 47    10     6  1666 1916      250 1326.1106 890.0245     150      100

#' @references Salvucci, D. D., & Goldberg, J. H. (2000). Identifying fixations and saccades in eye-tracking protocols. Proceedings of the Symposium on Eye Tracking Research & Applications - ETRA ’00, 71–78.